### 2025-07-31 #8677 @editorialbot
[PRE REVIEW]: MieSimulatorGUI: A user-friendly tool to compute and visualize light scattering by spherical dielectric particles
**Submitting author:** @janakarana (<a href="http://orcid.org/0000-0002-6069-6527">Janaka Ranasinghesagara</a>)
**Repository:** https://github.com/VirtualPhotonics/MieSimulatorGUI
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.3
