### 2025-07-31 e827f0a
Add MIT license file
