### 2025-07-31 #8677 @kyleniemeyer
@janakarana a few initial questions:

- I can see installation instructions using packages/executables, but what about building from source? At minimum I would expect this to be an option, but couldn't find any instructions or a script here.
- Is there a test suite? Our review criteria requires (at minimum) some way to objectively test the software, with an automated test suite being preferred.
- While the paper builds properly, it looks like the references aren't being compiled.
