### 2025-07-31 f45cb0d
Corrected superscript and subscript format of paper.md
