+++ paper/paper.md
@@ -50,41 +50,41 @@ Mie theory, derived from Maxwell's equations, provides a comprehensive framework
 
 # Main Features
 
-The `MieSimulatorGUI` employs the `BHMIE` code `[@Bohren1983]` and procedures described by `[@Wiscombe1979]` to simulate light scattering by spherical dielectric particles. This user-friendly tool calculates spectral dependencies of key optical properties including scattering, extinction and back scattering cross sections, scattering coefficient, reduced scattering coefficient, scattering amplitude matrix entries (S1, S2), phase function, average cosine of the phase function, and forward / backward scattering percentage `[@VandeHulst1957; @Bohren1983]`. These calculations can be performed for both monodisperse and polydisperse particle distributions `[@Gelebart1996]`. Moreover, the tool facilitates parameter estimation by fitting reduced scattering coefficient curves, a technique of significant value in tissue optics`[@Jacques2013]`. All relevant documentation for this tool is available on the [Mie Simulator GUI Wiki](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) page.
+The `MieSimulatorGUI` employs the `BHMIE` code `[@Bohren1983]` and procedures described by `[@Wiscombe1979]` to simulate light scattering by spherical dielectric particles. This user-friendly tool calculates spectral dependencies of key optical properties including scattering, extinction and back scattering cross sections, scattering coefficient, reduced scattering coefficient, scattering amplitude matrix entries ($\text{S}_1$, $\text{S}_2$), phase function, average cosine of the phase function, and forward / backward scattering percentage `[@VandeHulst1957; @Bohren1983]`. These calculations can be performed for both monodisperse and polydisperse particle distributions `[@Gelebart1996]`. Moreover, the tool facilitates parameter estimation by fitting reduced scattering coefficient curves, a technique of significant value in tissue optics`[@Jacques2013]`. All relevant documentation for this tool is available on the [Mie Simulator GUI Wiki](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) page.
 
 # Design and Functionality
 
-The `MieSimulatorGUI` is a cross-platform GUI application developed using the C/C++ programming language within the [Qt](https://www.qt.io/) framework. The GUI contains six interactive panels as shown in Fig. \autoref{fig:Figure1}. 
+The `MieSimulatorGUI` is a cross-platform GUI application developed using the C/C++ programming language within the [Qt](https://www.qt.io/) framework. The GUI contains six interactive panels as shown in \autoref{fig:Figure1}. 
 
 ![Figure 1: Six interactive panels of `MieSimulatorGUI`: (a) Input selection, (b) Particle size distribution, (c) Scattering coefficient, (d) Phase function, (e) Reduced Scattering,  and (f) Scattering Asymmetry \label{fig:Figure1}](Figure1.png)
 Figure 1: Six interactive panels of `MieSimulatorGUI`: (a) Input selection, (b) Particle size distribution, (c) Scattering coefficient, (d) Phase function, (e) Reduced Scattering, and (f) Scattering Asymmetry
 
 ## Input Selection Panel
 
[…]
