### 2025-07-31 c2db404
Update paper.bib
