+++ paper/paper.bib
@@ -1,3 +1,4 @@
+% BIBLIOGRAPHY -----------------------------------------
 
 @book{Bohren1983,
     address = {New York},
