### 2025-07-31 902af9a
Update paper.bib
