+++ paper/paper.bib
@@ -1,136 +1,126 @@
 % BIBLIOGRAPHY -----------------------------------------
 
 @book{Bohren1983,
-    address = {New York},
-    author = {Bohren, C F and Huffman, D R},
-    publisher = {John Wiley and Sons, Inc.},
-    title = {{Absorption and Scattering of Light by Small Particles}},
-    year = {1983}
+	title        = {{Absorption and Scattering of Light by Small Particles}},
+	author       = {Bohren, C F and Huffman, D R},
+	year         = 1983,
+	publisher    = {John Wiley and Sons, Inc.},
+	address      = {New York}
 }
-
 @article{Gelebart1996,
-    author = {G{\'{e}}l{\'{e}}bart, Bernard and Tinet, Eric and Tualle, Jean Michel and Avrillier, Sigrid},
-    doi = {10.1088/0963-9659/5/4/005},
-    issn = {0963-9659},
-    journal = {Pure and Applied Optics},
-    number = {4},
-    pages = {377--388},
-    title = {{Phase function simulation in tissue phantoms: a fractal approach}},
-    volume = {5},
-    year = {1996}
+	title        = {{Phase function simulation in tissue phantoms: a fractal approach}},
+	author       = {G{\'{e}}l{\'{e}}bart, Bernard and Tinet, Eric and Tualle, Jean Michel and Avrillier, Sigrid},
+	year         = 1996,
+	journal      = {Pure and Applied Optics},
+	volume       = 5,
+	number       = 4,
+	pages        = {377--388},
+	doi          = {10.1088/0963-9659/5/4/005},
+	issn         = {0963-9659}
 }
-
 @book{Goody1989,
-    address = {Oxford},
-    author = {Goody, R. M. and Yung, Y. L.},
-    edition = {2nd Edition},
-    pages = {544},
-    publisher = {Oxford University Press},
-    title = {{Atmospheric Radiation: Theoretical Basis}},
-    year = {1989}
+	title        = {{Atmospheric Radiation: Theoretical Basis}},
+	author       = {Goody, R. M. and Yung, Y. L.},
+	year         = 1989,
+	publisher    = {Oxford University Press},
+	address      = {Oxford},
+	pages        = 544,
+	edition      = {2nd Edition}
 }
-
 @article{Horvath2009,
-    author = {Horvath, Helmuth},
-    doi = {10.1016/j.jqsrt.2009.03.001},
-    issn = {00224073},
-    journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
-    number = {11},
-    pages = {783--786},
-    title = {{Light scattering: Mie and More commemorating 100 years of Mie's 1908 publication}},
-    volume = {110},
-    year = {2009}
+	title        = {{Light scattering: Mie and More commemorating 100 years of Mie's 1908 publication}},
+	author       = {Horvath, Helmuth},
+	year         = 2009,
+	journal      = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+	volume       = 110,
+	number       = 11,
+	pages        = {783--786},
+	doi          = {10.1016/j.jqsrt.2009.03.001},
+	issn         = {00224073}
 }
-
 @article{Jacques2013,
-    author = {Jacques, Steven L.},
-    doi = {10.1088/0031-9155/58/14/5007},
-    issn = {0031-9155},
-    journal = {Phys. Med. Biol.},
-    month = {jul},
-    number = {14},
-    pages = {5007--5008},
-    title = {{Optical properties of biological tissues: a review}},
-    url = {http://stacks.iop.org/0031-9155/58/i=14/a=5007?key=crossref.c531ac37cfa1f77bbc0d33a8e92de8c9},
-    volume = {58},
-    year = {2013}
[…]
