### 2025-07-31 421037e
Update paper.md
