+++ paper/paper.md
@@ -42,15 +42,15 @@ bibliography: paper.bib
 
 # Summary
 
-Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Mie theory predictions are used in a wide variety of scientific and engineering fields spanning fields such as atmospheric optics, particle characterization, computer graphics, nanofluids, remote sensing, and biomedical optics. We developed the `MieSimulatorGUI` to provide a user-friendly interface for users to perform and visualize computations of optical scattering by particles using Mie theory. By utilizing a C/C++ computational engine with the Qt framework, our cross-platform tool calculates key optical properties for instance scattering coefficients, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. With six interactive panels, the GUI enables users to specify input parameters, visualize particle distributions, and analyze scattering phenomena, including the fitting of spectrally-varying reduced scattering coefficient parameters, which is particularly valuable in fields such as tissue optics. The `MieSimulatorGUI` is open source, hosted on [GitHub](https://github.com/VirtualPhotonics/MieSimulatorGUI), available from a [download page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Downloads) and distributed under permissive [MIT license](https://opensource.org/license/mit).
+Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Mie theory predictions are used in a wide variety of scientific and engineering fields spanning fields such as atmospheric optics, particle characterization, computer graphics, nanofluids, remote sensing, and biomedical optics. We developed the `MieSimulatorGUI` to provide a user-friendly interface for users to perform and visualize computations of optical scattering by particles using Mie theory. By utilizing a C/C++ computational engine with the Qt framework, our cross-platform tool calculates key optical properties for instance scattering coefficients, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. With six interactive panels, the GUI enables users to specify input parameters, visualize particle distributions, and analyze scattering phenomena, including the fitting of spectrally-varying reduced scattering coefficient parameters, which is particularly valuable in fields such as tissue optics. The `MieSimulatorGUI` is open source, hosted on [GitHub](https://github.com/VirtualPhotonics/MieSimulatorGUI), available from a [download page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Downloads) and distributed under [the MIT license](https://opensource.org/license/mit).
 
 # Statement of need
 
[…]
