### 2025-08-01 2787498
Update LICENSE.MD
