### 2025-08-01 38588c4
Update LICENSE.MD
