### 2025-08-01 edfc384
Update LICENSE.MD
