### 2025-08-01 ed20730
Format License.MD
