+++ LICENSE.MD
@@ -1,10 +1,10 @@
-## License
+# License
+-------------------------------------
 
 
+## The MIT License (MIT)
 
-### The MIT License (MIT)
-
-##### Copyright (c) 2025 Virtual Photonics Technology Initiative
+#### Copyright (c) 2025 Virtual Photonics Technology Initiative
 
 _Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:_
 
@@ -14,19 +14,20 @@ _THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPL
 
 
 -------------------------------------
-### GNU LESSER GENERAL PUBLIC LICENSE (LGPLv3) for Qt
+## GNU LESSER GENERAL PUBLIC LICENSE (LGPLv3) for Qt
+
 
 GNU LESSER GENERAL PUBLIC LICENSE
 
 Version 3, 29 June 2007
 
-##### Copyright © 2007 Free Software Foundation, Inc. <https://fsf.org/>
+#### Copyright © 2007 Free Software Foundation, Inc. <https://fsf.org/>
 
 Everyone is permitted to copy and distribute verbatim copies of this license document, but changing it is not allowed.
 
 This version of the GNU Lesser General Public License incorporates the terms and conditions of version 3 of the GNU General Public License, supplemented by the additional permissions listed below.
 
-###### 0. Additional Definitions.
+##### 0. Additional Definitions.
 As used herein, “this License” refers to version 3 of the GNU Lesser General Public License, and the “GNU GPL” refers to version 3 of the GNU General Public License.
 
 “The Library” refers to a covered work governed by this License, other than an Application or a Combined Work as defined below.
@@ -39,35 +40,36 @@ The “Minimal Corresponding Source” for a Combined Work means the Correspondi
 
 The “Corresponding Application Code” for a Combined Work means the object code and/or source code for the Application, including any data and utility programs needed for reproducing the Combined Work from the Application, but excluding the System Libraries of the Combined Work.
 
-###### 1. Exception to Section 3 of the GNU GPL.
+##### 1. Exception to Section 3 of the GNU GPL.
 You may convey a covered work under sections 3 and 4 of this License without being bound by section 3 of the GNU GPL.
 
-###### 2. Conveying Modified Versions.
+##### 2. Conveying Modified Versions.
 If you modify a copy of the Library, and, in your modifications, a facility refers to a function or data to be supplied by an Application that uses the facility (other than as an argument passed when the facility is invoked), then you may convey a copy of the modified version:
-
+	
 a) under this License, provided that you make a good faith effort to ensure that, in the event an Application does not supply the function or data, the facility still operates, and performs whatever part of its purpose remains meaningful, or
 b) under the GNU GPL, with none of the additional permissions of this License applicable to that copy.
-###### 3. Object Code Incorporating Material from Library Header Files.
+##### 3. Object Code Incorporating Material from Library Header Files.
 The object code form of an Application may incorporate material from a header file that is part of the Library. You may convey such object code under terms of your choice, provided that, if the incorporated material is not limited to numerical parameters, data structure layouts and accessors, or small macros, inline functions and templates (ten or fewer lines in length), you do both of the following:
 
 a) Give prominent notice with each copy of the object code that the Library is used in it and that the Library and its use are covered by this License.
 b) Accompany the object code with a copy of the GNU GPL and this license document.
-###### 4. Combined Works.
+##### 4. Combined Works.
[…]
