### 2025-08-01 4aa2a2e
GLPLv3 license for QT
