### 2025-08-06 b39820f
Moved lib files to lib folder. Changed 'a' in power law to 'A'. Added default parameters
