+++ src/calculate.cpp
@@ -1,5 +1,6 @@
 #include "calculate.h"
-
+#include "miesimulation.h"
+#include "utilities.h"
 
 calculate::calculate()
 {
@@ -22,6 +23,7 @@ void calculate::DoSimulation(Ui_MainWindow *ui, parameters *para)
     double refRelRe = 1.0;
     double refRelIm = 0.0;
     double xPara = 0.0;
+    double stepTheta = (maxTheta - minTheta)/static_cast<double>(para->nTheta - 1);
 
     std::complex<double> *curS1 = new std::complex<double> [para->nTheta];
     std::complex<double> *curS2 = new std::complex<double> [para->nTheta];
@@ -70,7 +72,7 @@ void calculate::DoSimulation(Ui_MainWindow *ui, parameters *para)
             {
                 for (unsigned int t = 0; t < para->nTheta; t++)
                 {
-                    mu = cos(para->minTheta + t * para->stepTheta);
+                    mu = cos(minTheta + t * stepTheta);
                     sim.FarFieldSolutionForRealRefIndex(&cS1, &cS2, &qSca, &qExt, &qBack, xPara, refRelRe, mu);
                     curS1[t] = cS1;
                     curS2[t] = cS2;
@@ -80,7 +82,7 @@ void calculate::DoSimulation(Ui_MainWindow *ui, parameters *para)
             {
                 for (unsigned int t = 0; t < para->nTheta; t++)
                 {
-                    mu = cos(para->minTheta + t * para->stepTheta);
+                    mu = cos(minTheta + t * stepTheta);
                     sim.FarFieldSolutionForComplexRefIndex(&cS1, &cS2, &qSca, &qExt, &qBack, xPara,
                                                            std::complex<double>(refRelRe,-refRelIm), mu);  //multiply by -1 to use "n-ik" convention
                     curS1[t] = cS1;
@@ -156,6 +158,7 @@ void calculate::ComputeMuspAtRefWavel(parameters *para)
     double refRelRe = 1.0;
     double refRelIm = 0.0;
     double xPara = 0.0;
+    double stepTheta = (maxTheta - minTheta)/static_cast<double>(para->nTheta - 1);
 
     std::complex<double> *curS1 = new std::complex<double> [para->nTheta];
     std::complex<double> *curS2 = new std::complex<double> [para->nTheta];
@@ -182,7 +185,7 @@ void calculate::ComputeMuspAtRefWavel(parameters *para)
             {
                 for (unsigned int t = 0; t < para->nTheta; t++)
                 {
-                    mu = cos(para->minTheta + t * para->stepTheta);
+                    mu = cos(minTheta + t * stepTheta);
                     sim.FarFieldSolutionForRealRefIndex(&cS1, &cS2, &qSca, &qExt, &qBack, xPara, refRelRe, mu);
                     curS1[t] = cS1;
                     curS2[t] = cS2;
@@ -192,7 +195,7 @@ void calculate::ComputeMuspAtRefWavel(parameters *para)
             {
                 for (unsigned int t = 0; t < para->nTheta; t++)
                 {
-                    mu = cos(para->minTheta + t * para->stepTheta);
+                    mu = cos(minTheta + t * stepTheta);
                     sim.FarFieldSolutionForComplexRefIndex(&cS1, &cS2, &qSca, &qExt, &qBack, xPara,
                                                            std::complex<double>(refRelRe,-refRelIm), mu);  //multiply by -1 to use "n-ik" convention
                     curS1[t] = cS1;
[…]
