+++ src/lib/qwt/qwt_polar_grid.cpp
@@ -1,0 +1,1138 @@
+/* -*- mode: C++ ; c-file-style: "stroustrup" -*- *****************************
+ * QwtPolar Widget Library
+ * Copyright (C) 2008   Uwe Rathmann
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the Qwt License, Version 1.0
+ *****************************************************************************/
+
+#include "qwt_polar_grid.h"
+#include "qwt_painter.h"
+#include "qwt_text.h"
+#include "qwt_clipper.h"
+#include "qwt_scale_map.h"
+#include "qwt_scale_engine.h"
+#include "qwt_scale_div.h"
+#include "qwt_scale_draw.h"
+#include "qwt_round_scale_draw.h"
+#include <qpainter.h>
+#include <qpen.h>
+#include <float.h>
+
+static inline bool isClose( double value1, double value2 )
+{
+    return qAbs( value1 - value2 ) < DBL_EPSILON;
+}
+
+class QwtPolarGrid::AxisData
+{
+public:
+    AxisData():
+        isVisible( false ),
+        scaleDraw( NULL )
+    {
+    }
+    ~AxisData()
+    {
+        delete scaleDraw;
+    }
+
+    bool isVisible;
+    mutable QwtAbstractScaleDraw *scaleDraw;
+    QPen pen;
+    QFont font;
+};
+
+class QwtPolarGrid::GridData
+{
+public:
+    GridData():
+        isVisible( true ),
+        isMinorVisible( false )
+    {
+    }
+
+    bool isVisible;
+    bool isMinorVisible;
+    QwtScaleDiv scaleDiv;
+
+    QPen majorPen;
+    QPen minorPen;
+};
+
+class QwtPolarGrid::PrivateData
+{
+public:
+    GridData gridData[QwtPolar::ScaleCount];
+    AxisData axisData[QwtPolar::AxesCount];
+    QwtPolarGrid::DisplayFlags displayFlags;
+    QwtPolarGrid::GridAttributes attributes;
+};
+
+/*!
+   \brief Constructor
+
+   Enables major and disables minor grid lines.
+   The azimuth and right radial axis are visible. all other axes
+   are hidden. Autoscaling is enabled.
+*/
+QwtPolarGrid::QwtPolarGrid():
+    QwtPolarItem( QwtText( "Grid" ) )
+{
+    d_data = new PrivateData;
+
+    for ( int axisId = 0; axisId < QwtPolar::AxesCount; axisId++ )
+    {
+        AxisData &axis = d_data->axisData[axisId];
+        switch( axisId )
+        {
+            case QwtPolar::AxisAzimuth:
+            {
+                axis.scaleDraw = new QwtRoundScaleDraw;
+                axis.scaleDraw->setTickLength( QwtScaleDiv::MinorTick, 2 );
+                axis.scaleDraw->setTickLength( QwtScaleDiv::MediumTick, 2 );
+                axis.scaleDraw->setTickLength( QwtScaleDiv::MajorTick, 4 );
+                axis.isVisible = true;
+                break;
+            }
+            case QwtPolar::AxisLeft:
+            {
+                QwtScaleDraw *scaleDraw = new QwtScaleDraw;
+                scaleDraw->setAlignment( QwtScaleDraw::BottomScale );
+
+                axis.scaleDraw = scaleDraw;
+                axis.isVisible = false;
+                break;
+            }
+            case QwtPolar::AxisRight:
+            {
+                QwtScaleDraw *scaleDraw = new QwtScaleDraw;
+                scaleDraw->setAlignment( QwtScaleDraw::BottomScale );
+
[…]
