+++ src/lib/qcustomplot.cpp
@@ -25,7 +25,6 @@
 
 #include "qcustomplot.h"
 
-
 /* including file 'src/vector2d.cpp', size 7340                              */
 /* commit ce344b3f96a62e5f652585e55f1ae7c7883cd45b 2018-06-25 01:03:39 +0200 */
 

[…]
