+++ src/app/dialog/mainwindow.cpp
@@ -1,0 +1,925 @@
+/**********************************************************************
+** All MainWindow functions are listed in this file.
+**********************************************************************/
+
+#include "mainwindow.h"
+#include "lib/qwt/qwt_polar_curve.h"
+#include "calc/calculate.h"
+#include "calc/utilities.h"
+#include "dialog/plotdata.h"
+#include "dialog/optionsdialog.h"
+#include "dialog/displaydialog.h"
+#include "dialog/mainwindowsupport.h"
+
+
+/***********************Starting Main window and initializations ******************************/
+MainWindow::MainWindow(QWidget *parent) :
+    QMainWindow(parent), ui(new Ui::MainWindow)
+{
+    //Initialize UI
+    ui->setupUi(this);
+
+   //Intialize Window
+    Initialize();
+
+    //Connect signals and slots
+    connect(ui->customPlot_Csca, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotScatteringCrossSection(QMouseEvent*)));
+    connect(ui->customPlot_Cext, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotExtinctionCrossSection(QMouseEvent*)));
+    connect(ui->customPlot_Cback, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotBackscatteringCrossSection(QMouseEvent*)));
+    connect(ui->customPlot_SizePara, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotSizeParameter(QMouseEvent*)));
+    connect(ui->customPlot_Mus, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotMus(QMouseEvent*)));
+    connect(ui->customPlot_Musp, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotMusp(QMouseEvent*)));
+    connect(ui->customPlot_G, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotG(QMouseEvent*)));
+    connect(ui->customPlot_FB, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotForwardBackward(QMouseEvent*)));
+    connect(ui->customPlot_S1S2, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotS1S2(QMouseEvent*)));
+    connect(ui->customPlot_PhaseFunctionLinear, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotPhaseFunctionLinear(QMouseEvent*)));
+    connect(ui->customPlot_Distribution, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotDistribution(QMouseEvent*)));
+    connect(ui->customPlot_MuspPowerLaw, SIGNAL(mouseMove(QMouseEvent*)), SLOT(MouseOverPlotMuspPowerLaw(QMouseEvent*)));
+}
+
+MainWindow::~MainWindow()
+{
+    delete ui;
+}
+
+/***************************** Intialize *********************************************/
+void MainWindow::Initialize()
+{
+    //Initialize paramaters
+    mOtherPlotsFlag = false;
+    mDistPlotFlag = false;
+    mArrayFlag = false;
+    mLoadCustomNoGoodFlag = false;
+    mPara = new parameters();
+    mPara->polarCurve = new QwtPolarCurve();
+
+    MainWindowSupport support;
+    support.InitializeGUI(ui, mPara);
+    support.SetWidgets(ui);
+
+    PlotData plot;
+    plot.InitializeDistributionPlot(ui);
[…]
