+++ src/lib/qwt/qwt_plot_curve.h
@@ -1,337 +1,0 @@
-/* -*- mode: C++ ; c-file-style: "stroustrup" -*- *****************************
- * Qwt Widget Library
- * Copyright (C) 1997   Josef Wilgen
- * Copyright (C) 2002   Uwe Rathmann
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the Qwt License, Version 1.0
- *****************************************************************************/
-
-#ifndef QWT_PLOT_CURVE_H
-#define QWT_PLOT_CURVE_H
-
-#include "qwt_global.h"
-#include "qwt_plot_seriesitem.h"
-#include "qwt_series_data.h"
-#include "qwt_text.h"
-#include <qpen.h>
-#include <qstring.h>
-
-class QPainter;
-class QPolygonF;
-class QwtScaleMap;
-class QwtSymbol;
-class QwtCurveFitter;
-
-/*!
-  \brief A plot item, that represents a series of points
-
-  A curve is the representation of a series of points in the x-y plane.
-  It supports different display styles, interpolation ( f.e. spline )
-  and symbols.
-
-  \par Usage
-  <dl><dt>a) Assign curve properties</dt>
-  <dd>When a curve is created, it is configured to draw black solid lines
-  with in QwtPlotCurve::Lines style and no symbols. 
-  You can change this by calling
-  setPen(), setStyle() and setSymbol().</dd>
-  <dt>b) Connect/Assign data.</dt>
-  <dd>QwtPlotCurve gets its points using a QwtSeriesData object offering
-  a bridge to the real storage of the points ( like QAbstractItemModel ).
-  There are several convenience classes derived from QwtSeriesData, that also store
-  the points inside ( like QStandardItemModel ). QwtPlotCurve also offers
-  a couple of variations of setSamples(), that build QwtSeriesData objects from
-  arrays internally.</dd>
-  <dt>c) Attach the curve to a plot</dt>
-  <dd>See QwtPlotItem::attach()
-  </dd></dl>
-
-  \par Example:
-  see examples/bode
-
-  \sa QwtPointSeriesData, QwtSymbol, QwtScaleMap
-*/
-class QWT_EXPORT QwtPlotCurve: 
-    public QwtPlotSeriesItem, public QwtSeriesStore<QPointF>
-{
-public:
-    /*!
-        Curve styles.
-        \sa setStyle(), style()
-    */
-    enum CurveStyle
-    {
-        /*!
-           Don't draw a curve. Note: This doesn't affect the symbols.
-        */
-        NoCurve = -1,
-
-        /*!
-           Connect the points with straight lines. The lines might
-           be interpolated depending on the 'Fitted' attribute. Curve
-           fitting can be configured using setCurveFitter().
-        */
-        Lines,
-
-        /*!
-           Draw vertical or horizontal sticks ( depending on the 
-           orientation() ) from a baseline which is defined by setBaseline().
-        */
-        Sticks,
-
-        /*!
-           Connect the points with a step function. The step function
-           is drawn from the left to the right or vice versa,
-           depending on the QwtPlotCurve::Inverted attribute.
-        */
-        Steps,
-
-        /*!
-           Draw dots at the locations of the data points. Note:
-           This is different from a dotted line (see setPen()), and faster
-           as a curve in QwtPlotCurve::NoStyle style and a symbol 
-           painting a point.
-        */
-        Dots,
-
-        /*!
-           Styles >= QwtPlotCurve::UserCurve are reserved for derived
-           classes of QwtPlotCurve that overload drawCurve() with
-           additional application specific curve types.
-        */
-        UserCurve = 100
-    };
-
-    /*!
[…]
