### 2025-08-07 d960d46
Added GPL license for Qt
