### 2025-08-07 9e1ca18
Add tab spacing when saving pahse function data; remove the dependency of the main ui on the child ui
