+++ src/app/dialog/optionsdialog.h
@@ -2,7 +2,9 @@
 #define OPTIONSDIALOG_H
 
 #include <QDialog>
-#include "ui_mainwindow.h"
+#include <QRadioButton>
+#include <QComboBox>
+#include <QSlider>
 #include "ui_optionsdialog.h"
 #include "parameters.h"
 
@@ -18,7 +20,16 @@ public:
     explicit OptionsDialog(QWidget *parent = nullptr);
     ~OptionsDialog();
 
-    void SaveData(Ui_MainWindow *ui, parameters *para);
+    void SaveData(QRadioButton *radioButton_MonoDisperse,
+                  QRadioButton *radioButton_PolyDisperse,
+                  QRadioButton *radioButton_NumDen,
+                  QRadioButton *radioButton_VolFrac,
+                  QComboBox *comboBox_Distribution,
+                   QSlider *slider_ConcPercentChange,
+                  QRadioButton *radioButton_PhaseAverage,
+                  QRadioButton *radioButton_PhasePara,
+                  QRadioButton *radioButton_PhasePerp,
+                  parameters *para);
 
 private slots:
     void on_pushButton_ScatPara_clicked();
@@ -34,8 +45,17 @@ private:
     bool flagS2;
     Ui::OptionsDialog *ui;
 
-    void SaveScatPara(Ui_MainWindow *ui, parameters *para, QString fileName, double margin);
-    void SavePhaseFunction(Ui_MainWindow *ui, parameters *para, QString fileName);
+    void SaveScatPara(QRadioButton *radioButton_MonoDisperse,
+                      QRadioButton *radioButton_PolyDisperse,
+                      QRadioButton *radioButton_NumDen,
+                      QRadioButton *radioButton_VolFrac,
+                      QComboBox *comboBox_Distribution,
+                      parameters *para, QString fileName,
+                      double margin);
+    void SavePhaseFunction(QRadioButton *radioButton_PhaseAverage,
+                           QRadioButton *radioButton_PhasePara,
+                           QRadioButton *radioButton_PhasePerp,
+                           parameters *para, QString fileName);
     void SaveS1(parameters *para, QString fileName);
     void SaveS2(parameters *para, QString fileName);
     void RememberLastDirectory(QString fileName);
