+++ src/app/dialog/displaydialog.h
@@ -2,7 +2,9 @@
 #define DISPLAYDIALOG_H
 
 #include <QDialog>
-#include "ui_mainwindow.h"
+#include <QRadioButton>
+#include <QComboBox>
+#include <QSlider>
 #include "parameters.h"
 
 namespace Ui {
@@ -16,7 +18,16 @@ class DisplayDialog : public QDialog
 public:
     explicit DisplayDialog(QWidget *parent = 0);
     ~DisplayDialog();
-    void DisplayData(Ui_MainWindow *, parameters *);
+    void DisplayData(QRadioButton *radioButton_MonoDisperse,
+                     QRadioButton *radioButton_PolyDisperse,
+                     QRadioButton *radioButton_NumDen,
+                     QRadioButton *radioButton_VolFrac,
+                     QComboBox *comboBox_Distribution,
+                     QSlider *slider_ConcPercentChange,
+                     QSlider *slider_WL_PFPolar,
+                     QRadioButton *radioButton_PhaseAverage,
+                     QRadioButton *radioButton_PhasePara,
+                     QRadioButton *radioButton_PhasePerp, parameters *);
     Ui::DisplayDialog *ui;
 
 private slots:
