### 2025-08-08 1cbda46
Moved parameter sanity check from mainwindowsupport.cpp to parameters.cpp
