+++ src/app/parameters.cpp
@@ -4,6 +4,9 @@
 //**********************************************************************
 
 #include "parameters.h"
+#include "qcombobox.h"
+#include "qradiobutton.h"
+#include <QMessageBox>
 
 parameters::parameters(void)
 {
@@ -11,4 +14,205 @@ parameters::parameters(void)
 
 parameters::~parameters(void)
 {
+}
+
+//Check the validity of common parameters
+bool parameters::CheckCommonParameters(QRadioButton *radioButton_MonoDisperse,
+                                       QRadioButton *radioButton_NumDen,
+                                       QRadioButton *radioButton_VolFrac)
+{
+    QMessageBox msgBox, msgBoxWarn;
+    msgBox.setWindowTitle("Error");
+    msgBox.setIcon(QMessageBox::Critical);
+    msgBoxWarn.setWindowTitle("Warning");
+    msgBoxWarn.setIcon(QMessageBox::Warning);
+
+    if ((scatRefReal <= 0.0) || (medRef <= 0.0))
+    {
+        msgBox.setText("Refractive index cannot be zero");
+        msgBox.exec();
+        return 1;
+    }
+    if ((scatRefReal/medRef == 1.0))
+    {
+        msgBox.setText("Relative refractive index cannot be 1.0");
+        msgBox.exec();
+        return 1;
+    }
+    double m = scatRefReal / medRef;
+    if ((m < 0.05) || (m > 5.0))
+    {
+        msgBoxWarn.setText("Unrealistic relative refractive index.");
+        msgBoxWarn.setInformativeText("Check sphere and medium refractive index values");
+        msgBoxWarn.exec();
+        return 1;
+    }
+    if (scatRefImag >= 5.0)
+    {
+        msgBox.setText("Imaginary refractive index must be negative or less than 5.0.");
+        msgBox.exec();
+        return 1;
+    }
+    //Avoid zeros
+    if ((startWavel <= 0.0) || (endWavel <= 0.0))
+    {
+        msgBox.setText("The starting or ending wavelength cannot be zero");
+        msgBox.exec();
+        return 1;
+    }
+    //Avoid zeros
+    if (stepWavel <= 0.0)
+    {
+        msgBox.setText("Wavelength step cannot be zero");
+        msgBox.exec();
+        return 1;
+    }
+    if (startWavel < 50.0)
+    {
+        msgBox.setText("Minimum wavlength is 50nm");
+        msgBox.exec();
+        return 1;
+    }
+    if (endWavel > 3000.0)
+    {
+        msgBox.setText("Maximum wavlength is 3000nm");
+        msgBox.exec();
+        return 1;
+    }
+    if (endWavel - startWavel < 0.0)
+    {
+        msgBox.setText("The starting wavelength is greater than the ending wavelength");
+        msgBox.exec();
+        return 1;
+    }
+    if (radioButton_NumDen->isChecked())
+    {
+        if (sphNumDensity <= 0.0)
+        {
+            msgBox.setText("Sphere concentration cannot be zero");
+            msgBox.exec();
+            return 1;
+        }
+    }
+    if (radioButton_VolFrac->isChecked())
+    {
+        if (volFraction <= 0.0)
+        {
+            msgBox.setText("Volume Fraction cannot be zero");
+            msgBox.exec();
+            return 1;
+        }
+
+        if (volFraction >= 1.0)
+        {
+            msgBox.setText("Volume Fraction cannot exceed 1.0");
+            msgBox.exec();
+            return 1;
+        }
+    }
+    if ((meanRadius < 0.00005) ||(meanRadius >150))
+    {
+        msgBox.setText("Diameter is out of range!");
+        msgBox.setInformativeText("Enter a value between 0.0001μm and 300μm");
[…]
