+++ src/app/calc/parameters.cpp
@@ -1,14 +1,0 @@
-//**********************************************************************
-//** "parameters" variable is used to store all variables in a single
-//** variable. This allows easy portability.
-//**********************************************************************
-
-#include "parameters.h"
-
-parameters::parameters(void)
-{
-}
-
-parameters::~parameters(void)
-{
-}

[…]
