### 2025-08-08 2821d59
Corrected forward and backward percentage calculations
