### 2025-08-09 7526288
Added unit tests for the functions in calculate.cpp
