+++ src/test/test_calculate.cpp
@@ -1,0 +1,359 @@
+#include <QDebug>
+#include <QtMath>
+#include <complex>
+
+#include "test_calculate.h"
+
+QTEST_MAIN(TestCalculate)
+
+TestCalculate::TestCalculate()
+{
+}
+
+TestCalculate::~TestCalculate()
+{
+}
+
+//Initialize values for tests
+void TestCalculate::init()
+{
+    mPara = new parameters();
+    mCalc = new calculate();
+
+    mPara->radArray = new double[mPara->nRadius];
+    mPara->numDensityArray = new double[mPara->nRadius];
+    mPara->scatRefRealArray = new double[mPara->nRadius];
+    mPara->scatRefImagArray = new double[mPara->nRadius];
+    mPara->wavelArray = new double[mPara->nWavel];
+    mPara->mus = new double[mPara->nWavel];
+    mPara->g = new double[mPara->nWavel];
+    mPara->cSca = new double[mPara->nWavel];
+    mPara->cExt = new double[mPara->nWavel];
+    mPara->cBack = new double[mPara->nWavel];
+    mPara->SizePara = new double[mPara->nWavel];
+    mPara->forward = new double[mPara->nWavel];
+    mPara->backward = new double[mPara->nWavel];
+
+    mPara->S1 = new std::complex<double>*[mPara->nWavel];
+    mPara->S2 = new std::complex<double>*[mPara->nWavel];
+    mPara->phaseFunctionAve = new double*[mPara->nWavel];
+    mPara->phaseFunctionPara = new double*[mPara->nWavel];
+    mPara->phaseFunctionPerp = new double*[mPara->nWavel];
+
+    for (unsigned int w = 0; w < mPara->nWavel; ++w)
+    {
+        mPara->S1[w] = new std::complex<double>[mPara->nTheta];
+        mPara->S2[w] = new std::complex<double>[mPara->nTheta];
+        mPara->phaseFunctionAve[w] = new double[mPara->nTheta];
+        mPara->phaseFunctionPara[w] = new double[mPara->nTheta];
+        mPara->phaseFunctionPerp[w] = new double[mPara->nTheta];
+    }
+}
+
+// clean up resources
+void TestCalculate::cleanup()
+{
+    if (mPara->S1) {
+        for (unsigned int w = 0; w < mPara->nWavel; ++w)
+            delete[] mPara->S1[w];
+        delete[] mPara->S1;
+        mPara->S1 = nullptr;
+    }
+    if (mPara->S2) {
+        for (unsigned int w = 0; w < mPara->nWavel; ++w)
+            delete[] mPara->S2[w];
+        delete[] mPara->S2;
+        mPara->S2 = nullptr;
+    }
+    if (mPara->phaseFunctionAve) {
+        for (unsigned int w = 0; w < mPara->nWavel; ++w)
+            delete[] mPara->phaseFunctionAve[w];
+        delete[] mPara->phaseFunctionAve;
+        mPara->phaseFunctionAve = nullptr;
+    }
+    if (mPara->phaseFunctionPara) {
+        for (unsigned int w = 0; w < mPara->nWavel; ++w)
+            delete[] mPara->phaseFunctionPara[w];
+        delete[] mPara->phaseFunctionPara;
+        mPara->phaseFunctionPara = nullptr;
+    }
+    if (mPara->phaseFunctionPerp) {
+        for (unsigned int w = 0; w < mPara->nWavel; ++w)
+            delete[] mPara->phaseFunctionPerp[w];
+        delete[] mPara->phaseFunctionPerp;
[…]
