### 2025-08-11 14013a8
Seperated sanity checks and eroor mesaages
