+++ src/app/parameters.cpp
@@ -4,8 +4,8 @@
 //**********************************************************************
 
 #include "parameters.h"
-#include "qcombobox.h"
-#include "qradiobutton.h"
+#include <QComboBox>
+#include <QRadioButton>
 #include <QMessageBox>
 
 parameters::parameters(void)
@@ -20,199 +20,235 @@ parameters::~parameters(void)
 bool parameters::CheckCommonParameters(QRadioButton *radioButton_MonoDisperse,
                                        QRadioButton *radioButton_NumDen,
                                        QRadioButton *radioButton_VolFrac)
+{    
+    bool monoDisperseSelection = false;
+    bool numDenSelection = false;
+    bool volFracSelection = false;
+
+    if (radioButton_MonoDisperse->isChecked())
+        monoDisperseSelection = true;
+    if (radioButton_NumDen->isChecked())
+        numDenSelection = true;
+    if (radioButton_VolFrac->isChecked())
+        volFracSelection = true;
+
+    ParameterValidationResult check = CheckValidityCommonParameters(monoDisperseSelection,
+                                                         numDenSelection,volFracSelection);
+    if (!check.isValid)
+    {
+        QMessageBox msgBoxError;
+        msgBoxError.setWindowTitle("Error");
+        msgBoxError.setIcon(QMessageBox::Critical);
+        msgBoxError.setText(check.errorMessage);
+        msgBoxError.exec();
+        return false;
+    }
+    else
+        return true;
+}
+
+ParameterValidationResult parameters::CheckValidityCommonParameters(
+                                                    bool monoDisperseSelection,
+                                                    bool numDenSelection,
+                                                    bool volFracSelection)
 {
-    QMessageBox msgBox, msgBoxWarn;
-    msgBox.setWindowTitle("Error");
-    msgBox.setIcon(QMessageBox::Critical);
-    msgBoxWarn.setWindowTitle("Warning");
-    msgBoxWarn.setIcon(QMessageBox::Warning);
+    ParameterValidationResult result;
+    result.isValid = true;
 
     if ((scatRefReal <= 0.0) || (medRef <= 0.0))
     {
-        msgBox.setText("Refractive index cannot be zero");
-        msgBox.exec();
-        return 1;
+        result.isValid = false;
+        result.errorMessage = "Refractive index cannot be zero.";
+        return result;
+
     }
     if ((scatRefReal/medRef == 1.0))
     {
-        msgBox.setText("Relative refractive index cannot be 1.0");
-        msgBox.exec();
-        return 1;
+        result.isValid = false;
+        result.errorMessage = "Relative refractive index cannot be 1.0.";
+        return result;
     }
     double m = scatRefReal / medRef;
     if ((m < 0.05) || (m > 5.0))
     {
-        msgBoxWarn.setText("Unrealistic relative refractive index.");
-        msgBoxWarn.setInformativeText("Check sphere and medium refractive index values");
-        msgBoxWarn.exec();
-        return 1;
+        result.isValid = false;
+        result.errorMessage = "Unrealistic relative refractive index! Check sphere and medium refractive index values.";
+        return result;
     }
     if (scatRefImag >= 5.0)
     {
-        msgBox.setText("Imaginary refractive index must be negative or less than 5.0.");
-        msgBox.exec();
-        return 1;
+        result.isValid = false;
+        result.errorMessage = "Imaginary refractive index must be negative or less than 5.0.";
+        return result;
     }
     //Avoid zeros
     if ((startWavel <= 0.0) || (endWavel <= 0.0))
     {
-        msgBox.setText("The starting or ending wavelength cannot be zero");
-        msgBox.exec();
-        return 1;
+        result.isValid = false;
[…]
