+++ src/app/dialog/mainwindow.cpp
@@ -87,7 +87,7 @@ void MainWindow::on_pushButton_RunSimulation_clicked()
     else
     {
 
-        if (!mPara->CheckCommonParameters(ui->radioButton_MonoDisperse,
+        if (mPara->CheckCommonParameters(ui->radioButton_MonoDisperse,
                                           ui->radioButton_NumDen,
                                           ui->radioButton_VolFrac))
         {
@@ -113,7 +113,7 @@ void MainWindow::on_pushButton_RunSimulation_clicked()
             //Poly disperse
             if (ui->radioButton_PolyDisperse->isChecked())
             {
-                if (!mPara->CheckDistributionParameters(ui->comboBox_Distribution))   //sanity check
+                if (mPara->CheckDistributionParameters(ui->comboBox_Distribution))   //sanity check
                 {
                     support.ProcessDistribution(ui, mPara, static_cast<unsigned int>(ui->comboBox_Distribution->currentIndex()));
                     support.ProcessPolyDisperse(ui,mPara);
@@ -162,11 +162,11 @@ void MainWindow::on_pushButton_ShowDistributionAndCustom_clicked()
     else
     {
         support.LoadInputData(ui, mPara);
-        if (!mPara->CheckCommonParameters(ui->radioButton_MonoDisperse,
+        if (mPara->CheckCommonParameters(ui->radioButton_MonoDisperse,
                                           ui->radioButton_NumDen,
                                           ui->radioButton_VolFrac))    //sanity check
         {
-            if (!mPara->CheckDistributionParameters(ui->comboBox_Distribution))   //sanity check
+            if (mPara->CheckDistributionParameters(ui->comboBox_Distribution))   //sanity check
             {
                 support.SetWidgets(ui);
                 support.ProcessDistribution(ui, mPara, static_cast<unsigned int>(ui->comboBox_Distribution->currentIndex()));
