### 2025-08-12 e6bad6a
Added script to app.pro file to build the test project before building the app. Fixed missing tab in optionsdialog.
