### 2025-08-15 fa92b4a
Introduced enums for polydisperse distribution types and reference wavelengths; Changed qwtSlider to QSlider
