### 2025-08-19 e0895f3
Added tick calculation function and unit tests
