+++ src/test/test_utilities.h
@@ -20,6 +20,10 @@ private slots:
     void test_ComplexAbsSquared();
     void test_ComplexAbs();
     void test_SimpsonsWeight();
+    void test_NiceStep_SmallRange();
+    void test_NiceStep_LargeRange();
+    void test_NiceStep_EdgeCases();
+
 
 private:
     utilities *mUtilities;
