### 2025-08-22 615f7b0
Remove the qwt library dependency; Upgrade to Qt 6; Addd new polar plot for phase function; Add the option to save induvidual plots; Fix naming conventions; add two new sample files; Add new unit tests
