+++ src/app/calc/utilities.cpp
@@ -4,16 +4,16 @@
 
 #include "calc/utilities.h"
 
-utilities::utilities()
+Utilities::Utilities()
 {
 }
 
-utilities::~utilities()
+Utilities::~Utilities()
 {
 }
 
 //Delay function for display update
-void utilities::Delay()
+void Utilities::Delay()
 {
     QTime dieTime= QTime::currentTime().addMSecs(1);
     while( QTime::currentTime() < dieTime )
@@ -21,26 +21,27 @@ void utilities::Delay()
 }
 
 //Intensity calculation (Amplitude^2)
-double utilities::ComplexAbsSquared(std::complex<double> c)
+double Utilities::ComplexAbsSquared(std::complex<double> c)
 {
     return (c.real()*c.real() + c.imag()*c.imag());
 }
 
 //Absolute value (amplitude) calculation
-double utilities::ComplexAbs(std::complex<double> c)
+double Utilities::ComplexAbs(std::complex<double> c)
 {
     return sqrt(ComplexAbsSquared(c));
 }
 
 //Simpson's 1/3 rule
-double utilities::SimpsonsWeight (unsigned int i, unsigned int n)
+double Utilities::SimpsonsWeight (unsigned int i, unsigned int n)
 {
     if (i == 0 || i == n - 1)  return 1.0 / 3.0;
     if (i % 2 != 0)   return 4.0 / 3.0;
     return 2.0 / 3.0;
 }
 
-double utilities::NiceStep(double range, int initialCircles)
+//Find nice step for polar plot ticks
+double Utilities::NiceStep(double range, int initialCircles)
 {
     double niceStep = 0.0;
     double roughStep = (initialCircles > 0) ? range / initialCircles : range;
@@ -70,4 +71,20 @@ double utilities::NiceStep(double range, int initialCircles)
         }
     }
     return niceStep;
+}
+
+//Find min and max out of three variables
+double Utilities::FindMinMax(const QVector<double>& yPara, const QVector<double>& yPerp,
+                             const QVector<double>& yAve, bool flagMinMax)
+{
+    QVector<double> allY;
+    allY << yPara << yPerp << yAve;
+
+    if (allY.isEmpty())
+        return std::nan("");
+
+    if (flagMinMax)
+        return *std::min_element(allY.constBegin(), allY.constEnd());
+    else
+        return *std::max_element(allY.constBegin(), allY.constEnd());
 }
