+++ src/app/dialog/plotdata.cpp
@@ -3,12 +3,8 @@
 **********************************************************************/
 
 #include "plotdata.h"
+#include "calc/utilities.h"
 #include "lib/qcustomplot.h"
-#include "lib/qwt/qwt_polar_curve.h"
-#include "lib/qwt/qwt_scale_engine.h"
-
-const QwtInterval radialInterval( 0.0, 10.0 );
-const QwtInterval azimuthInterval( 0.0, 360.0 );
 
 PlotData::PlotData(void)
 {
@@ -18,153 +14,156 @@ PlotData::~PlotData(void)
 {
 }
 
+//Clear and Initialize Plots
+void PlotData::ClearPlots(Ui_MainWindow *ui)
+{
+    InitialSetupDistributionPlot(ui);
+    InitialSetupPhaseFunctionPolarPlot(ui);
+    InitialSetupPhaseFunctionLinearPlot(ui);
+    InitialSetupS1S2Plot(ui);
+    InitialSetupMuspPowerLawFit(ui);
+    InitialSetupOtherPlots(ui);
+}
+
 //Initialize distribution plot
-void PlotData::InitializeDistributionPlot(Ui_MainWindow *ui)
+void PlotData::InitialSetupDistributionPlot(Ui_MainWindow *ui)
 {
     double minX, maxX;
     minX = 0.95 * (ui->lineEdit_Diameter->text().toDouble());
     maxX = 1.05 * (ui->lineEdit_Diameter->text().toDouble());
 
-    ui->customPlot_Distribution->xAxis->setLabel("Diameter (μm)");
-    ui->customPlot_Distribution->xAxis->setRange(minX, maxX);
-    ui->customPlot_Distribution->yAxis->setLabel("Ns  (/vol. of 1mm³)");
-    ui->customPlot_Distribution->clearPlottables();
-    ui->customPlot_Distribution->replot();
+    auto customPlot = ui->customPlot_Distribution;
+
+    customPlot->xAxis->setLabel("Diameter (μm)");
+    customPlot->xAxis->setRange(minX, maxX);
+    customPlot->yAxis->setLabel("Ns  (/vol. of 1mm³)");
+    customPlot->yAxis->setRange(0, 5);
+    RemovePlotables(customPlot);
+    customPlot->replot();
 }
 
-//Initialize polar plot
-void PlotData::InitializePhaseFunctionPolarPlot(Ui_MainWindow *ui, parameters *para )
+//Initialize phase function linear plot
+void PlotData::InitialSetupPhaseFunctionLinearPlot(Ui_MainWindow *ui)
 {
-//    SetSliderWL(ui);
+    auto customPlot = ui->customPlot_PhaseFunctionLinear;
 
-    //Polar plot: phase function
-    ui->qwtpolarplot_PhaseFunctionPolar->setAutoReplot( false );
-    ui->qwtpolarplot_PhaseFunctionPolar->setPlotBackground( Qt::white );
-    ui->qwtpolarplot_PhaseFunctionPolar->setScale( QwtPolar::Azimuth,0, 360, 360 / 12 );
-    ui->qwtpolarplot_PhaseFunctionPolar->setScale( QwtPolar::Radius, 0, 1);
-    mGrid = new QwtPolarGrid();
-    mGrid->setPen( QPen( Qt::black ) );
-    for ( int scaleId = 0; scaleId < QwtPolar::ScaleCount; scaleId++ )
-    {
-        mGrid->showGrid( scaleId );
-        mGrid->showMinorGrid( scaleId );
-        QPen minorPen( Qt::gray );
-        #if 0
-            minorPen.setStyle( Qt::DotLine );
-        #endif
[…]
