### 2025-08-22 3aaa69f
Change curPlot to customPlot
