### 2025-08-23 91055fc
Remove qwt library
