+++ src/app/lib/qwt/qwt_painter.cpp
@@ -1,1299 +1,0 @@
-/* -*- mode: C++ ; c-file-style: "stroustrup" -*- *****************************
- * Qwt Widget Library
- * Copyright (C) 1997   Josef Wilgen
- * Copyright (C) 2002   Uwe Rathmann
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the Qwt License, Version 1.0
- *****************************************************************************/
-
-#include "qwt_painter.h"
-#include "qpainterpath.h"
-#include "qwt_math.h"
-#include "qwt_clipper.h"
-#include "qwt_color_map.h"
-#include "qwt_scale_map.h"
-#include <qwindowdefs.h>
-#include <qwidget.h>
-#include <qframe.h>
-#include <qrect.h>
-#include <qpainter.h>
-#include <qpalette.h>
-#include <qpaintdevice.h>
-#include <qpixmap.h>
-#include <qstyle.h>
-#include <qtextdocument.h>
-#include <qabstracttextdocumentlayout.h>
-#include <qstyleoption.h>
-#include <qpaintengine.h>
-#include <qapplication.h>
-#include <qdesktopwidget.h>
-
-#if QT_VERSION >= 0x050000
-#include <qwindow.h>
-#endif
-
-#if QT_VERSION < 0x050000 
-
-#ifdef Q_WS_X11
-#include <qx11info_x11.h>
-#endif
-
-#endif
-
-bool QwtPainter::d_polylineSplitting = true;
-bool QwtPainter::d_roundingAlignment = true;
-
-static inline bool qwtIsClippingNeeded( 
-    const QPainter *painter, QRectF &clipRect )
-{
-    bool doClipping = false;
-    const QPaintEngine *pe = painter->paintEngine();
-    if ( pe && pe->type() == QPaintEngine::SVG )
-    {
-        // The SVG paint engine ignores any clipping,
-
-        if ( painter->hasClipping() )
-        {
-            doClipping = true;
-            clipRect = painter->clipRegion().boundingRect();
-        }
-    }
-
-    return doClipping;
-}
-
-template <class T>
-static inline void qwtDrawPolyline( QPainter *painter,
-    const T *points, int pointCount, bool polylineSplitting )
-{
-    bool doSplit = false;
-    if ( polylineSplitting )
-    {
-        const QPaintEngine *pe = painter->paintEngine();
-        if ( pe && pe->type() == QPaintEngine::Raster )
-        {
-            /*
-                The raster paint engine seems to use some algo with O(n*n).
-                ( Qt 4.3 is better than Qt 4.2, but remains unacceptable)
-                To work around this problem, we have to split the polygon into
-                smaller pieces.
-             */
-            doSplit = true;
-        }
-    }
-
-    if ( doSplit )
-    {
-        const int splitSize = 6;
-
-        for ( int i = 0; i < pointCount; i += splitSize )
-        {
-            const int n = qMin( splitSize + 1, pointCount - i );
-            painter->drawPolyline( points + i, n );
-        }
-    }
-    else
-    {
-        painter->drawPolyline( points, pointCount );
-    }
-}
-
-static inline QSize qwtScreenResolution()
-{
-    static QSize screenResolution;
-    if ( !screenResolution.isValid() )
-    {
-        QDesktopWidget *desktop = QApplication::desktop();
-        if ( desktop )
-        {
-            screenResolution.setWidth( desktop->logicalDpiX() );
-            screenResolution.setHeight( desktop->logicalDpiY() );
-        }
-    }
-
-    return screenResolution;
-}
-
[…]
