+++ src/app/lib/qwt/qwt_polar_layout.cpp
@@ -1,444 +1,0 @@
-/* -*- mode: C++ ; c-file-style: "stroustrup" -*- *****************************
- * QwtPolar Widget Library
- * Copyright (C) 2008   Uwe Rathmann
- *
- * This library is free software; you can redistribute it and/or
- * modify it under the terms of the Qwt License, Version 1.0
- *****************************************************************************/
-
-#include "qwt_polar_layout.h"
-#include "qwt_polar_plot.h"
-#include "qwt_polar_canvas.h"
-#include "qwt_text.h"
-#include "qwt_text_label.h"
-#include "qwt_legend.h"
-#include <qscrollbar.h>
-
-class QwtPolarLayout::LayoutData
-{
-public:
-    void init( const QwtPolarPlot *, const QRectF &rect );
-
-    struct t_legendData
-    {
-        int frameWidth;
-        int hScrollExtent;
-        int vScrollExtent;
-        QSizeF hint;
-    } legend;
-
-    struct t_titleData
-    {
-        QwtText text;
-        int frameWidth;
-    } title;
-
-    struct t_canvasData
-    {
-        int frameWidth;
-    } canvas;
-};
-
-void QwtPolarLayout::LayoutData::init(
-    const QwtPolarPlot *plot, const QRectF &rect )
-{
-    // legend
-
-    if ( plot->plotLayout()->legendPosition() != QwtPolarPlot::ExternalLegend
-            && plot->legend() )
-    {
-        legend.frameWidth = plot->legend()->frameWidth();
-        legend.hScrollExtent =
-            plot->legend()->scrollExtent( Qt::Horizontal );
-        legend.vScrollExtent =
-            plot->legend()->scrollExtent( Qt::Vertical );
-
-        const QSizeF hint = plot->legend()->sizeHint();
-
-        double w = qMin( hint.width(), rect.width() );
-        double h = plot->legend()->heightForWidth( w );
-        if ( h == 0.0 )
-            h = hint.height();
-
-        if ( h > rect.height() )
-            w += legend.hScrollExtent;
-
-        legend.hint = QSizeF( w, h );
-    }
-
-    // title
-
-    title.frameWidth = 0;
-    title.text = QwtText();
-
-    if ( plot->titleLabel() )
-    {
-        const QwtTextLabel *label = plot->titleLabel();
-        title.text = label->text();
-        if ( !( title.text.testPaintAttribute( QwtText::PaintUsingTextFont ) ) )
-            title.text.setFont( label->font() );
-
-        title.frameWidth = plot->titleLabel()->frameWidth();
-    }
-
-    // canvas
-
-    canvas.frameWidth = plot->canvas()->frameWidth();
-}
-
-class QwtPolarLayout::PrivateData
-{
-public:
-    PrivateData():
-        margin( 0 ),
-        spacing( 0 )
-    {
-    }
-
-    QRectF titleRect;
-    QRectF legendRect;
-    QRectF canvasRect;
-
-    QwtPolarLayout::LayoutData layoutData;
-
-    QwtPolarPlot::LegendPosition legendPos;
-    double legendRatio;
-
-    unsigned int margin;
-    unsigned int spacing;
-};
-
-/*!
-  \brief Constructor
- */
-
-QwtPolarLayout::QwtPolarLayout()
-{
-    d_data = new PrivateData;
-
-    setLegendPosition( QwtPolarPlot::BottomLegend );
-    invalidate();
-}
-
-//! Destructor
-QwtPolarLayout::~QwtPolarLayout()
-{
-    delete d_data;
-}
-
-/*!
-  \brief Specify the position of the legend
-  \param pos The legend's position.
-  \param ratio Ratio between legend and the bounding rect
[…]

[…]
