### 2025-08-25 fe93668
Update paper.md
