### 2025-08-31 0e7a92b
Add radio buttons to plot a phase function with all polarizations, S1 and S2 curves; Rename curPlot to customPlot; Update banner
