+++ src/app/dialog/plotdata.cpp
@@ -28,48 +28,34 @@ void PlotData::ClearPlots(Ui_MainWindow *ui)
 //Initialize distribution plot
 void PlotData::InitialSetupDistributionPlot(Ui_MainWindow *ui)
 {
+    const double BUFFER_PERCENT = 0.05;
     double minX, maxX;
-    minX = 0.95 * (ui->lineEdit_Diameter->text().toDouble());
-    maxX = 1.05 * (ui->lineEdit_Diameter->text().toDouble());
+    minX = (1 - BUFFER_PERCENT) * (ui->lineEdit_Diameter->text().toDouble());
+    maxX = (1 + BUFFER_PERCENT) * (ui->lineEdit_Diameter->text().toDouble());
 
     auto customPlot = ui->customPlot_Distribution;
 
-    customPlot->xAxis->setLabel("Diameter (μm)");
-    customPlot->xAxis->setRange(minX, maxX);
-    customPlot->yAxis->setLabel("Ns  (/vol. of 1mm³)");
-    customPlot->yAxis->setRange(0, 5);
     RemovePlotables(customPlot);
-    customPlot->replot();
+    InitialSetupPlot(customPlot, "Diameter (μm)", "Ns  (/vol. of 1mm³)", minX, maxX);
 }
 
 //Initialize phase function linear plot
 void PlotData::InitialSetupPhaseFunctionLinearPlot(Ui_MainWindow *ui)
-{
+{    
+    double minX = -180;
+    double maxX = 180;
+
     auto customPlot = ui->customPlot_PhaseFunctionLinear;
+    InitialSetupPlot(customPlot, "Angle (deg.)", "Magnitude", minX, maxX);
 
-    // remove previous graphs (if any)
-    RemoveGraphs(customPlot);
-
-    // remove previous items (if any)
-    RemoveItems(customPlot);
-
-    // remove previous legends (if any)
-    RemoveLegends(customPlot);
-
-    customPlot->xAxis->setLabel("Angle (deg.)");
-    customPlot->xAxis->setRange(-180,180);
     QVector<double> ticks{-180, -135, -90, -45, 0, 45, 90, 135, 180};
     QVector<QString> labels;
     for (double tick : ticks)
         labels << QString::number(tick);
-
     QSharedPointer<QCPAxisTickerText> textTicker(new QCPAxisTickerText);
     textTicker->addTicks(ticks, labels);
     customPlot->xAxis->setTicker(textTicker);
-    customPlot->yAxis->setLabel("Magnitude");
-    customPlot->yAxis->setRange(0, 5);
     customPlot->replot();
-    customPlot->legend->setVisible(false);
 }
 
 //Initialize phase function polar plot
@@ -79,12 +65,14 @@ void PlotData::InitialSetupPhaseFunctionPolarPlot(Ui_MainWindow *ui)
     mPolarMinRadius = 0.0;
     mPolarMaxRadius = 1.0;
     DrawPolarPlotGrid(customPlot, true); //true for linear
+    customPlot->setInteractions(QCP::iRangeDrag | QCP::iRangeZoom);
 }
 
-//Initialize S1/S2
+//Initialize S1,S2
 void PlotData::InitialSetupS1S2Plot(Ui_MainWindow *ui)
 {
     auto customPlot = ui->customPlot_S1S2;
+    customPlot->setInteractions(QCP::iRangeDrag | QCP::iRangeZoom);
 
     // remove previous graphs (if any)
     RemoveGraphs(customPlot);
@@ -109,92 +97,27 @@ void PlotData::InitialSetupOtherPlots(Ui_MainWindow *ui)
     double minX = ui->lineEdit_StartWL->text().toDouble();
[…]
