+++ src/app/dialog/mainwindow.cpp
@@ -119,7 +119,7 @@ void MainWindow::UpdateMuspFitErrorDisplay()
 {
     ui->label_CurrentMSE->setText("<font color=\"red\">M.S. Error = "
                                   "</font>"+QString::number(mPara->muspFittingError,'g',6));
-    ui->label_CurrentA->setText("<font color=\"blue\">A = </font>"+
+    ui->label_CurrentA->setText("A = "+
                                 QString::number(mPara->muspAtRefWavel[mPara->refWavelIdx],'g',6)+
                                 " mm<sup>-1</sup>");
 }
@@ -554,6 +554,12 @@ void MainWindow::on_radioButton_S2_clicked()
     UpdateS1S2Plot();
 }
 
+//radioButtonS2_clicked: Select S1/S2 plot
+void MainWindow::on_radioButton_S1S2_clicked()
+{
+    UpdateS1S2Plot();
+}
+
 //radioButton_S1S2_Abs_clicked: Select absolute value
 void MainWindow::on_radioButton_S1S2_Abs_clicked()
 {
@@ -602,6 +608,12 @@ void MainWindow::on_radioButton_PhasePerp_clicked()
     UpdatePhaseFunctionPolarPlot();
 }
 
+//radioButton_PhasePerp_clicked: Select both Parallel and Perpendicular components
+void MainWindow::on_radioButton_PhaseAll_clicked()
+{
+    UpdatePhaseFunctionPolarPlot();
+}
+
 //radioButton_dTheta0_1 is clicked: Simulation time warning
 void MainWindow::on_radioButton_Phase_DTheta0_1_clicked()
 {
@@ -640,7 +652,7 @@ void MainWindow::on_radioButton_FittingSimple_clicked()
     ui->labelF_0_8->setDisabled(true);
     ui->labelF_1_0->setDisabled(true);
     ui->doubleSpinBox_F->setDisabled(true);
-    QString labelB = "<P><b><FONT COLOR='#aa0000'>";
+    QString labelB = "<P><b><FONT COLOR='#ff5500'>";
     labelB .append("b");
     labelB .append("</b></P></br>");
     ui->label_BLabel->setText(labelB);
@@ -659,11 +671,11 @@ void MainWindow::on_radioButton_FittingComplex_clicked()
     ui->labelF_0_6->setDisabled(false);
     ui->labelF_0_8->setDisabled(false);
     ui->labelF_1_0->setDisabled(false);
-    QString labelB = "<P><b><FONT COLOR='#aa0000'>";
+    QString labelB = "<P><b><FONT COLOR='#ff5500'>";
     labelB .append("b<sub>Mie</sub>");
     labelB .append("</b></P></br>");
     ui->label_BLabel->setText(labelB);
-    QString labelF = "<P><b><FONT COLOR='#005500'>";
+    QString labelF = "<P><b><FONT COLOR='#00c800'>";
     labelF .append("f<sub>Ray</sub>");
     labelF .append("</b></P></br>");
     ui->label_FLabel->setText(labelF);
@@ -796,13 +808,13 @@ void MainWindow::on_comboBox_Distribution_currentIndexChanged(int value)
 void MainWindow::MouseOverPlotScatteringCrossSection(QMouseEvent *event)
 {
     QString strNameY;
-    QCustomPlot *curPlot = ui->customPlot_Csca;
+    QCustomPlot *customPlot = ui->customPlot_Csca;
     QString strNameX = "WL";
     if (ui->radioButton_LogYAxis->isChecked())
         strNameY = "Log(Csca)";
     else
         strNameY = "Csca";
-    DisplayGraphData(event, curPlot, strNameX, strNameY);
+    DisplayGraphData(event, customPlot, strNameX, strNameY);
 }
 
 
[…]
