### 2025-09-04 5cccb9f
Update LICENSE.MD
