+++ LICENSE.MD
@@ -1,8 +1,4 @@
-# License
--------------------------------------
-
-
-## The MIT License (MIT)
+## The MIT License
 
 #### Copyright (c) 2025 Virtual Photonics Technology Initiative
 
