### 2025-09-04 86e7f35
Update LICENSE.MD
