### 2025-09-04 6e1def4
Update LICENSE.MD
