### 2025-09-04 698c209
Update LICENSE.MD
