### 2025-11-14 #8677 @janakarana
Hi @kyleniemeyer

This manuscript was submitted over four months ago and has remained on the waiting list since September 29th.
We would appreciate it if you could provide an update on its current status and assign an editor soon.

To expedite the process, we respectfully suggest the following potential Topic Editors, based on their individual research profiles, for your consideration:
Henrik Finsberg (@finsberg)
Rohit Goswami (@HaoZeke)
Mojtaba Barzegari (@mbarzegary)

We also suggest the following individuals as potential reviewers:
Scott Prahl (scottprahl)
Kyle Beggs (kylebeggs)
Joshua Beirer (jbeirer)

Thanks
