### 2025-11-19 #8677 @finsberg
@kyleniemeyer, sure I can take on this submission (although this is quite outside my area of expertise).
