### 2025-11-29 #8677 @finsberg
Hi @zmoon @scottprahl @jbeirer @cpinte
Considering your expertise, would any of you be willing to review this submission? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

If you are not available it would be great if you have any suggestions for potential reviewers.

Paper: https://github.com/openjournals/joss-papers/blob/joss.08677/joss.08677/10.21105.joss.08677.pdf
Software: https://github.com/VirtualPhotonics/MieSimulatorGUI
