### 2025-12-04 #8677 @GiannakopoulosIlias
Hello,

Here is a review per the guidelines.

- Contribution and authorship: Please list the per author contribution in the acknowledgments.

- Installation:
In the ReadME or a separate documentation document please provide a clearly-stated list of dependencies. If not needed clarify that the installation is automatic.

- Community guidelines:
In the ReadME or a separate documentation document please provide guidelines for third parties wishing to contribute to the software, report issues or problems with the software, or seek support. For the last two, setting a contact email, or specifying for the users to use the issues page should be enough.

-The State of the field is missing:
The authors need to describe how this software compares to other Mie Series packages.
I have an open package, and I am aware of a few other available codes in Git. A few others can be found here: https://en.wikipedia.org/wiki/Codes_for_electromagnetic_scattering_by_spheres. I am not aware of any other package that has a UI.

-Quality of writing:
Typo: Line 51: Figure ??.

-References:
The list of references is not complete: I would cite the other packages and the original paper of Gustav Mie.
