### 2025-12-06 4fb6193
Update paper.bib
