+++ paper/paper.bib
@@ -1,5 +1,15 @@
 % BIBLIOGRAPHY -----------------------------------------
 
+@article{Mie1908,
+	author = {Mie, Gustav},
+  	title = {{Beiträge zur Optik trüber Medien, speziell kolloidaler Metallösungen}},
+  	journal = {Annalen der Physik},
+  	volume = {330},
+  	number = {3},
+  	pages = {377--445},
+  	year = {1908},
+  	doi = {10.1002/andp.19083300302}
+}
 @book{Bohren1983,
 	title        = {{Absorption and Scattering of Light by Small Particles}},
 	author       = {Bohren, C F and Huffman, D R},
