### 2025-12-06 38b6d94
Update paper.md
