### 2025-12-06 #8677 @janakarana
Dear @GiannakopoulosIlias and  @ovidiopr

No readme file is available for this software. The installation instructions for Windows, Linux and macOS are provided on the following page. https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki
