### 2025-12-06 #8677 @ayusmin
Dear @finsberg <https://github.com/finsberg>,

Thank you for the invitation.
I would like to review the paper.

Thanks and Regards,
Ayusmin

On Fri, Dec 5, 2025 at 1:38 AM Henrik Finsberg ***@***.***>
wrote:

> *finsberg* left a comment (openjournals/joss-reviews#8677)
> <https://github.com/openjournals/joss-reviews/issues/8677#issuecomment-3614137045>
>
> Hi @ovidiopr <https://github.com/ovidiopr> @cfinch
> <https://github.com/cfinch> @GiannakopoulosIlias
> <https://github.com/GiannakopoulosIlias> @ayusmin
> <https://github.com/ayusmin>
> Considering your expertise, would any of you be willing to review this
> submission? We carry out our checklist-driven reviews here in GitHub issues
> and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
> If you are not available it would be great if you have any suggestions for
> potential reviewers.
>
> Paper:
> https://github.com/openjournals/joss-papers/blob/joss.08677/joss.08677/10.21105.joss.08677.pdf
> Software: https://github.com/VirtualPhotonics/MieSimulatorGUI
>
