### 2025-12-06 #9551 @editorialbot
[REVIEW]: MieSimulatorGUI: A user-friendly tool to compute and visualize light scattering by spherical dielectric particles
**Submitting author:** @janakarana (<a href="http://orcid.org/0000-0002-6069-6527">Janaka Ranasinghesagara</a>)
**Repository:** https://github.com/VirtualPhotonics/MieSimulatorGUI
**Branch with paper.md** (empty if default branch): joss
**Version:** v2.0

@MartinPdeS & @GiannakopoulosIlias & @ayusmin & @ovidiopr, your review will be checklist based. Each of you will have a separate checklist that you should update when carrying out your review.
