+++ README.MD
@@ -1,7 +1,7 @@
 ## Mie Simulator GUI
 The [Mie Simulator GUI](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) is an open-source, user-friendly tool designed to calculate the characteristics of Mie scatterers. It determines various properties, including the spectral dependence of the scattering coefficient, scattering cross-section, reduced scattering coefficient, phase function, average cosine of the phase function, and forward/backward scattering percentage. Additionally, it calculates scattering matrix entries for single or multiple wavelengths. Based on the Lorentz-Mie solution, these calculations can be performed for both mono-disperse and poly-disperse distributions. The tool can also estimate fitting parameters from the reduced scattering coefficient curve, though its capabilities are limited to the characteristics of the far-field scattering pattern. 
 
-This GUI tool is based on the [BHMIE code](https://onlinelibrary.wiley.com/doi/book/10.1002/9783527618156). For absorbing spheres, the tool uses the complex refractive index definition (m<sub>sphere</sub> = m<sub>real</sub> – j m<sub>imag</sub>) as described in van de Hulst's [Light Scattering by Small Particles (1957) book](https://books.google.com/books?id=PlHfPMVAFRcC&source=gbs_navlinks_s). To understand the Mie algorithm and other parameters within the GUI, refer to, read Bohren and Huffman's [Absorption and Scattering of Light by Small Particles, (1983) book](https://onlinelibrary.wiley.com/doi/book/10.1002/9783527618156), Maetzler's [research report](http://www.atmo.arizona.edu/students/courselinks/spring09/atmo656b/maetzler_mie_v2.pdf) and Scott Prahl's ['Mie Basics'](https://miepython.readthedocs.io/en/latest/01_basics.html#) web page. Phase function calculations for mixtures of spheres are based on the work by Gélébart et al. in [Pure and Applied Optics, 5 (1996)](https://iopscience.iop.org/article/10.1088/0963-9659/5/4/005/meta). The reduced scattering power law fitting functions are derived from  Steve L. Jacques's review paper in [Phys. Med & Bio. 58 (2013)](https://iopscience.iop.org/article/10.1088/0031-9155/58/11/R37).
[…]
