### 2025-12-12 d555c1d
Revised manuscript and added new references
