+++ paper/paper.bib
@@ -1,14 +1,16 @@
 % BIBLIOGRAPHY -----------------------------------------
 
-@article{Mie1908,
-	author = {Mie, Gustav},
-  	title = {{Beiträge zur Optik trüber Medien, speziell kolloidaler Metallösungen}},
-  	journal = {Annalen der Physik},
-  	volume = {330},
-  	number = {3},
-  	pages = {377--445},
-  	year = {1908},
-  	doi = {10.1002/andp.19083300302}
+@article{Bhandari11,
+	title      = {Modeling optical properties of human skin using Mie theory for particles with different size distributions and refractive indices},
+	author     = {A. Bhandari and B. Hamre and {\O}. Frette and K. Stamnes and J. J. Stamnes},
+	journal    = {Opt. Express},	
+	number     = {15},
+	pages      = {14549--14567},
+	publisher  = {Optica Publishing Group},	
+	volume     = {19},
+	month = {Jul},
+	year = {2011},
+	doi = {10.1364/OE.19.014549}
 }
 @book{Bohren1983,
 	title        = {{Absorption and Scattering of Light by Small Particles}},
@@ -16,6 +18,17 @@
 	year         = 1983,
 	publisher    = {John Wiley and Sons, Inc.},
 	address      = {New York}
+}
+@article{Chalut2008,
+	title={Application of Mie theory to assess structure of spheroidal scattering in backscattering geometries},
+	author={Chalut, Kevin J and Giacomelli, Michael G and Wax, Adam},
+	journal={J Opt Soc Am A Opt Image Sci Vis},
+	volume={25},
+	number={8},
+	pages={1866--1874},
+	year={2008},
+	publisher={Optical Society of America},
+	doi={10.1364/josaa.25.001866}
 }
 @article{Gelebart1996,
 	title        = {{Phase function simulation in tissue phantoms: a fractal approach}},
@@ -73,6 +86,40 @@
 	issn         = {1559-128X},
 	pmid         = 32225380
 }
+@article{Mie1908,
+	author = {Mie, Gustav},
+  	title = {{Beiträge zur Optik trüber Medien, speziell kolloidaler Metallösungen}},
+  	journal = {Annalen der Physik},
+  	volume = {330},
+  	number = {3},
+  	pages = {377--445},
+  	year = {1908},
+  	doi = {10.1002/andp.19083300302}
+}
+@misc{MieScattering, 
+	author="{Giannakopoulos, Ilias}", 
+	title={Matlab software for the {M}ie Scattering of {PEC} and Homogeneous Spheres Irradiated by a Linearly Polarized Plane Wave}, 
+	url = {https://www.github.com/GiannakopoulosIlias/MieScattering}, 
+	year={2018} 
+}
+@article{PoinsinetdeSivry-Houle2023,
+    author = {Martin Poinsinet de Sivry-Houle and Nicolas Godbout and Caroline Boudoux},
+    journal = {Opt. Continuum},
+    title = {PyMieSim: an open-source library for fast and flexible far-field Mie scattering simulations},
+    volume = {2},
+    number = {3},
+    pages = {520--534},
+    year = {2023},
+    doi = {10.1364/OPTCON.473102},
+}
+@misc{Prahl_mie,
+	author = {Scott Prahl},
+	title = {Mie Scattering},
+	version = {v2.6.3},
+	year = {2023},
+	doi = {10.5281/zenodo.10087653},
+	url = {https://github.com/scottprahl/mie}
+}
 @article{Saidi1995,
 	title        = {{Mie and Rayleigh modeling of visible-light scattering in neonatal skin}},
 	author       = {Saidi, Iyad S. and Jacques, Steven L. and Tittel, Frank K.},
@@ -102,6 +149,17 @@
[…]
