### 2025-12-12 #9551 @janakarana
Dear @GiannakopoulosIlias
Thank you for agreeing to review the manuscript and providing initial comments. Our responses to your concerns are given below.

Contribution and authorship:
Please list the per author contribution in the acknowledgments.

**Response:** We have included each author's contribution in the acknowledgments.

**Action:** We have added the following text to Acknowledgements. “JCR was the primary contributor, handling the core software development, UI design, validation, GitHub upload, and drafting the original manuscript. VV provided overall supervision and critically reviewed and edited the final manuscript. CH wrote the initial Mie scattering code, and LG assisted with the GitHub upload.

Installation:
In the ReadME or a separate documentation document please provide a clearly-stated list of dependencies. If not needed clarify that the installation is automatic.

**Response:** Thank you for the comment. Installation instructions and dependency requirements for the source code on Windows, Linux, and macOS are detailed on separate wiki pages, which are referenced on the README.md page.

**Action:** To address the comment directly, a sentence in the manuscript has been revised to: "All relevant documentation, including a list of dependencies and complete installation guides for Windows, Linux, and macOS, is available on the [Mie Simulator GUI Wiki](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) page." Furthermore, we have confirmed that the README.md prominently links to the wiki pages detailing the dependencies.

Community guidelines:
In the ReadME or a separate documentation document please provide guidelines for third parties wishing to contribute to the software, report issues or problems with the software, or seek support. For the last two, setting a contact email, or specifying for the users to use the issues page should be enough.

**Response:** Guidelines for third parties wishing to contribute, including coding conventions and discussion forums, are detailed on the project's GitHub Wiki pages, with links provided on the README.md file. For reporting issues, problems, or seeking support, users are directed to use the GitHub Discussions page to ensure transparency and community tracking.

**Action:** Links for contributing to the software, reporting issues, and seeking support are now included in README.md.

State of the field:
Currently Missing!
The authors need to describe how this software compares to other Mie Series packages. I have an open package, and I am aware of a few other available codes in Git. A few others can be found here: https://en.wikipedia.org/wiki/Codes_for_electromagnetic_scattering_by_spheres. I am not aware of any other package that has a UI.

**Response:** We thank the reviewer for pointing out the need for a clearer comparison with other available Mie packages. We acknowledge the existence of numerous Mie scattering codes, as listed on resources like SCATTPORT.org and the Wikipedia page provided. While the 1000-word limit imposed by the journal restricts us by providing an exhaustive review, we have revised the ‘Statement of Need’ section of the manuscript to address this point and explicitly highlight the unique features of our package.

**Action:** A part of the ‘Statement of Need’ section currently reads as: “ …. While numerous Mie simulation packages exist, many of which are listed in resources like [SCATTPORT.org](https://scattport.org/) and [Wikipedia](https://en.wikipedia.org/wiki/Codes_for_electromagnetic_scattering_by_spheres), they generally fall into two categories: older, established codes focusing on computational efficiency [@Wiscombe1980; @Bohren1983], and newer, object-oriented libraries often found on Git repositories [@Sumlin2018; @PoinsinetdeSivry-Houle2023; @Prahl_mie; @MieScattering]. MieSimulatorGUI distinguishes itself from other actively developed, open-source Mie software packages by offering an intuitive, cross-platform, full desktop application-style GUI. It allows researchers, engineers and teachers to rapidly compute, visualize, and download critical metrics associated with scattering phenomena without needing to write any command, script, or code. By combining a powerful C/C++ computational engine with intuitive [Qt](https://www.qt.io/) graphical interface, the MieSimulatorGUI provides user-friendly access to complex Mie theory computations for streamlined analysis and educational purposes.”

Quality of writing:
Typo: Line 51: Figure ??.

**Response:** Thanks for bringing this to our attention.

**Action:**  Fixed as “Figure 1”

References:
The list of references is not complete: I would cite the other packages and the original paper of Gustav Mie.

**Response:** We have added the original paper by Gustav Mie and several new references to ensure the completeness of the list, including citations for other relevant packages and literature.

**Action:** We added a reference to the original paper by Gustav Mie (Mie1908). The following new references were added to the paper.bib file.
Bhandari11, Chalut2008, MieScattering, PoinsinetdeSivry-Houle2023, Prahl_mie, Sumlin2018, Wiscombe1980
