+++ .github/workflows/test.yml
@@ -14,8 +14,7 @@ jobs:
           version: '6.2.4' 
 
       - name: Build and Run Tests
-        run: |
-          mkdir build && cd build
+        run: 
           qmake ../MieSimulatorGUI.pro
           make -j$(nproc)
           ./test/MieSimulatorGUI_test
