+++ .github/workflows/test.yml
@@ -15,6 +15,8 @@ jobs:
 
       - name: Build and Run Tests
         run: 
+          mkdir build && cd build
           qmake ../MieSimulatorGUI.pro
           make -j$(nproc)
-          ./test/MieSimulatorGUI_test
+          cd test
+          ./MieSimulatorGUI_test
