+++ .github/workflows/test.yml
@@ -19,12 +19,18 @@ jobs:
         with:
           version: '6.2.4' 
 
-      - name: Create Build Directory, Run qmake and Build
+      - name: Change permission
+        run: chmod +x MieSimulatorGUI.pro
+        
+      - name: Create Build Directory
         run: mkdir -p build
-             chmod +x MieSimulatorGUI.pro
              cd build
-             qmake ../MieSimulatorGUI.pro             
-             make -j$(nproc)
+
+      - name: Run qmake
+        run: qmake ../MieSimulatorGUI.pro  
+             
+      - name: Build     
+        run: make -j$(nproc)
 
       - name: Run Tests
         working-directory: ./build/test
