### 2026-01-06 36e8cc5
Update paper and references
