+++ paper/paper.bib
@@ -1,5 +1,18 @@
 % BIBLIOGRAPHY -----------------------------------------
 
+@article{Aernouts2013,
+	author = {Ben Aernouts and Eduardo Zamora-Rojas and Robbe Van Beers and Rodrigo Watt\'{e} and Ling Wang and Mizuki Tsuta and Jeroen Lammertyn and Wouter Saeys},
+	journal = {Opt. Express},
+	number = {26},
+	pages = {32450--32467},
+	publisher = {Optica Publishing Group},
+	title = {Supercontinuum laser based optical characterization of Intralipid{\textregistered} phantoms in the 500-2250 nm range},
+	volume = {21},
+	month = {Dec},
+	year = {2013},
+	url = {https://opg.optica.org/oe/abstract.cfm?URI=oe-21-26-32450},
+	doi = {10.1364/OE.21.032450},
+}
 @article{Bhandari11,
 	title      = {Modeling optical properties of human skin using Mie theory for particles with different size distributions and refractive indices},
 	author     = {A. Bhandari and B. Hamre and {\O}. Frette and K. Stamnes and J. J. Stamnes},
@@ -29,6 +42,28 @@
 	year={2008},
 	publisher={Optical Society of America},
 	doi={10.1364/josaa.25.001866}
+}
+@article{DiNinni2011,
+	author = {Di Ninni, P and Martelli, F and Zaccanti, G},
+	title = {{Intralipid: towards a diffusive reference standard for optical tissue phantoms}},
+	journal = {Physics in Medicine \& Biology},
+	volume = {56},
+	number = {2},
+	pages = {N21--N28},
+	year = {2011},
+	publisher = {IOP Publishing},
+	doi = {10.1088/0031-9155/56/2/N01},  
+}
+@article{Galy2020,
+	title = {Revisiting independent versus dependent scattering regimes in suspensions or aggregates of spherical particles},
+	author = {Galy, Tiphaine and Huang, Daniel and Pilon, Laurent},
+	journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+	volume = {246},
+	pages = {106924},
+	year = {2020},
+	publisher = {Elsevier},
+	doi = {10.1016/j.jqsrt.2020.106924},
+	issn = {0022-4073},  
 }
 @article{Gelebart1996,
 	title        = {{Phase function simulation in tissue phantoms: a fractal approach}},
@@ -60,6 +95,17 @@
 	pages        = {783--786},
 	doi          = {10.1016/j.jqsrt.2009.03.001},
 	issn         = {00224073}
+}
+@article{Ivezic1996,
+	title = {An investigation of dependent/independent scattering regimes using a discrete dipole approximation},
+	journal = {International Journal of Heat and Mass Transfer},
+	author = {Zˇ. Ivezić and M. Pinar Mengüç},
+	volume = {39},
+	number = {4},
+	pages = {811-822},
+	year = {1996},
+	issn = {0017-9310},
+	doi = {https://doi.org/10.1016/0017-9310(95)00142-5},
 }
 @article{Jacques2013,
 	title        = {{Optical properties of biological tissues: a review}},
@@ -102,6 +148,20 @@
 	url = {https://www.github.com/GiannakopoulosIlias/MieScattering}, 
 	year={2018} 
 }
+@article{Mourant1997,
+	author = {Judith R. Mourant and Tamika Fuselier and James Boyer and Tamara M. Johnson and Irving J. Bigio},
+	journal = {Appl. Opt.},
+	keywords = {Absorption coefficient; Elastic scattering; Mie scattering; Optical properties; Refractive index; Turbid media},
+	number = {4},
+	pages = {949--957},
[…]
