+++ paper/paper.md
@@ -41,54 +41,60 @@ bibliography: paper.bib
 
 # Summary
 
-Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Mie theory predictions are used in a wide variety of scientific and engineering fields spanning fields such as atmospheric optics, particle characterization, computer graphics, nanofluids, remote sensing, and biomedical optics. We developed the `MieSimulatorGUI` to provide a user-friendly interface for users to perform and visualize computations of optical scattering by particles using Mie theory. By utilizing a C/C++ computational engine with the Qt framework, our cross-platform tool calculates key optical properties for instance scattering coefficients, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. With six interactive panels, the GUI enables users to specify input parameters, visualize particle distributions, and analyze scattering phenomena, including the fitting of spectrally-varying reduced scattering coefficient parameters, which is particularly valuable in fields such as tissue optics. The `MieSimulatorGUI` is open source, hosted on [GitHub](https://github.com/VirtualPhotonics/MieSimulatorGUI), available from a [download page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Downloads) and distributed under [the MIT license](https://opensource.org/license/mit).
+Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Its predictions are essential across various scientific and engineering disciplines, including biomedical optics, atmospheric optics, particle characterization, nanofluids, computer graphics, and remote sensing. To bridge the gap for researchers who require Mie simulations but lack specialized programming expertise, we developed `MieSimulatorGUI`. By integrating a high-performance C/C++ computational engine with the Qt framework, this user-friendly cross-platform tool calculates key optical properties such as scattering coefficients, cross-sections, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. The graphical user interface (GUI) features six interactive panels that allow users to specify input parameters, visualize distributions, and analyze scattering phenomena. Furthermore, it enables the fitting of spectrally varying reduced scattering coefficient parameters, a feature particularly valuable in fields like tissue optics.
+`MieSimulatorGUI` is open source, hosted on [GitHub](https://github.com/VirtualPhotonics/MieSimulatorGUI) and distributed under [the MIT license](https://opensource.org/license/mit); it is available via the [project's download page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Downloads).
 
 # Statement of need
 
-Mie theory, derived from Maxwell's equations, provides a comprehensive framework for modeling electromagnetic scattering by spherical particles [@Mie1908]. This theory is invaluable in fields ranging from nanomaterials and biomedical optics to atmospheric science and astronomy [@Goody1989; @Saidi1995; @Wang2005; @Chalut2008; @Horvath2009, @Bhandari11]. Despite its broad applicability, the theory's reliance on complex mathematical constructs, such as infinite series and special functions [@VandeHulst1957; @Bohren1983; @Wiscombe1980; @Majic2020], demands advanced computational implementations.
[…]
