### 2026-01-06 cf46cb5
Update paper and references
