+++ paper/paper.bib
@@ -105,7 +105,7 @@
 	pages = {811-822},
 	year = {1996},
 	issn = {0017-9310},
-	doi = {https://doi.org/10.1016/0017-9310(95)00142-5},
+	doi = {10.1016/0017-9310(95)00142-5},
 }
 @article{Jacques2013,
 	title        = {{Optical properties of biological tissues: a review}},
@@ -238,7 +238,7 @@
 	pages        = {127-134},
 	year         = {2018},
 	issn         = {0022-4073},
-	doi          = {https://doi.org/10.1016/j.jqsrt.2017.10.012},
+	doi          = {10.1016/j.jqsrt.2017.10.012},
 	publisher    = {Elsevier}
 }
 @Article{Teri2022,
@@ -249,7 +249,7 @@
 	year = {2022},
 	number = {10},
 	pages = {3161--3187},
-	doi = {https://doi.org/10.5194/amt-15-3161-2022}
+	doi = {10.5194/amt-15-3161-2022}
 }
 @book{VandeHulst1957,
 	title        = {{Light scattering by small particles}},
@@ -271,16 +271,6 @@
 	issn         = 10833668,
 	pmid         = 16292956
 }
-@article{Wiscombe1980,
-	author = {Wiscombe, W. J.},
-	title = {Improved Mie Scattering Calculations: MIEV0},
-	journal = {Applied Optics},
-	volume = {19},
-	issue = {9},
-	pages = {1505--1509},
-	year = {1980},
-	publisher = {Optical Society of America}
-}
 @techreport{Wiscombe1979,
 	title        = {{Mie Scattering Calculations : Advances in Technique and Fast, Vector-Speed Computer Codes}},
 	author       = {Wiscombe, Warren J},
@@ -292,4 +282,16 @@
 	institution  = {National Center for Atmospheric Research},
 	keywords     = {Mie Theory,Scattering,Sphere},
 	mendeley-tags = {Mie Theory,Scattering,Sphere}
+}
+@article{Wiscombe1980,
+	author = {W. J. Wiscombe},
+	journal = {Appl. Opt.},
+	number = {9},
+	pages = {1505--1509},
+	publisher = {Optica Publishing Group},
+	title = {Improved Mie scattering algorithms},
+	volume = {19},
+	month = {May},
+	year = {1980},
+	doi = {10.1364/AO.19.001505},
 }
