+++ paper/paper.md
@@ -46,7 +46,7 @@ Mie theory is a mathematical framework derived from Maxwell's equations that mod
 
 # Statement of need
 
-Mie theory, derived from Maxwell's equations, provides a comprehensive framework for modeling electromagnetic scattering by spherical particles [@Mie1908]. This theory is invaluable across diverse fields, ranging from nanomaterials and biomedical optics to atmospheric science and astronomy [@Goody1989; @Saidi1995; @Wang2005; @Chalut2008; @Horvath2009, @Bhandari11]. Despite its broad applicability, the theory's reliance on complex mathematical constructs, such as infinite series and special functions [@VandeHulst1957; @Bohren1983; @Wiscombe1980; @Majic2020], demands advanced computational implementation. 
+Mie theory, derived from Maxwell's equations, provides a comprehensive framework for modeling electromagnetic scattering by spherical particles [@Mie1908]. This theory is invaluable across diverse fields, ranging from nanomaterials and biomedical optics to atmospheric science and astronomy [@Goody1989; @Saidi1995; @Wang2005; @Chalut2008; @Horvath2009; @Bhandari11]. Despite its broad applicability, the theory's reliance on complex mathematical constructs, such as infinite series and special functions [@VandeHulst1957; @Bohren1983; @Wiscombe1980; @Majic2020], demands advanced computational implementation. 
 
 While numerous Mie simulation packages are avialble (many of which are listed on [SCATTPORT.org](https://scattport.org) and [Wikipedia](https://en.wikipedia.org/wiki/Codes_for_electromagnetic_scattering_by_spheres)), they generally fall into two categories: older, established codes focusing on computational efficiency [@Wiscombe1980; @Bohren1983], and newer, object-oriented libraries typically hosted on version-control platforms [@Sumlin2018; @PoinsinetdeSivry-Houle2023; @Prahl_mie; @MieScattering]. Although both categories provide robust computational engines, they usually demand significant programming proficiency. This requirement creates a barrier for experimentalists, clinical scientists, and educators who need these analytical capabilities but may lack the specialized coding expertise to integrate such libraries into their workflows.
 
@@ -56,12 +56,14 @@ While numerous Mie simulation packages are avialble (many of which are listed on
 
 Built on the BHMIE [@Bohren1983] and Wiscombe [@Wiscombe1979] frameworks, `MieSimulatorGUI` calculates spectral optical properties, including scattering coefficients, cross-sections, scattering amplitude matrix entries ($\text{S}_1$, $\text{S}_2$), phase functions, and scattering asymmetry. The tool supports both monodisperse and polydisperse particle distributions [@Gelebart1996] and facilitates parameter estimation by fitting reduced scattering coefficient curves, a technique of significant value in tissue optics [@Jacques2013]. To ensure computational accuracy and GUI stability, the software includes an automated test suite integrated via GitHub Actions. Its high-performance C++ engine enables near-instantaneous computation and plotting, providing real-time visual updates across the spectral range. Comprehensive documentation, including cross-platform installation guides (Windows, Linux, and macOS), dependency specifications, command-line test execution and several examples, is available on the [Mie Simulator GUI Wiki](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) page. 
 
[…]
