+++ paper/paper.md
@@ -48,7 +48,7 @@ Mie theory is a mathematical framework derived from Maxwell's equations that mod
 
 Mie theory, derived from Maxwell's equations, provides a comprehensive framework for modeling electromagnetic scattering by spherical particles [@Mie1908]. This theory is invaluable across diverse fields, ranging from nanomaterials and biomedical optics to atmospheric science and astronomy [@Goody1989; @Saidi1995; @Wang2005; @Chalut2008; @Horvath2009; @Bhandari11]. Despite its broad applicability, the theory's reliance on complex mathematical constructs, such as infinite series and special functions [@VandeHulst1957; @Bohren1983; @Wiscombe1980; @Majic2020], demands advanced computational implementation. 
 
-While numerous Mie simulation packages are avialble (many of which are listed on [SCATTPORT.org](https://scattport.org) and [Wikipedia](https://en.wikipedia.org/wiki/Codes_for_electromagnetic_scattering_by_spheres)), they generally fall into two categories: older, established codes focusing on computational efficiency [@Wiscombe1980; @Bohren1983], and newer, object-oriented libraries typically hosted on version-control platforms [@Sumlin2018; @PoinsinetdeSivry-Houle2023; @Prahl_mie; @MieScattering]. Although both categories provide robust computational engines, they usually demand significant programming proficiency. This requirement creates a barrier for experimentalists, clinical scientists, and educators who need these analytical capabilities but may lack the specialized coding expertise to integrate such libraries into their workflows.
+While numerous Mie simulation packages are available (many of which are listed on [SCATTPORT.org](https://scattport.org) and [Wikipedia](https://en.wikipedia.org/wiki/Codes_for_electromagnetic_scattering_by_spheres)), they generally fall into two categories: older, established codes focusing on computational efficiency [@Wiscombe1980; @Bohren1983], and newer, object-oriented libraries typically hosted on version-control platforms [@Sumlin2018; @PoinsinetdeSivry-Houle2023; @Prahl_mie; @MieScattering]. Although both categories provide robust computational engines, they usually demand significant programming proficiency. This requirement creates a barrier for experimentalists, clinical scientists, and educators who need these analytical capabilities but may lack the specialized coding expertise to integrate such libraries into their workflows.
 
 `MieSimulatorGUI` bridges this gap by providing an intuitive, cross-platform desktop application that computes and fits scattering parameters for monodisperse and polydisperse distributions without requiring any coding. Unlike standard implementations, it supports heterogeneous polydispersity, allowing users to assign bin-specific complex refractive indices via custom data inputs, a feature often absent in simplified GUI tools. The tool facilitates high-impact use cases such as biomedical optics [@Mourant1997; @Wang2005; @Jacques2013] and atmospheric research [@Seinfeld1998; @Teri2022], where users can define complex configurations and directly fit spectrally varying reduced scattering coefficients. By integrating a powerful C/C++ computational engine with intuitive [Qt](https://www.qt.io/) interface, `MieSimulatorGUI` offers accessible yet powerful Mie theory computations, facilitating both streamlined research analysis and interactive pedagogical demonstrations. 
 
@@ -57,7 +57,7 @@ While numerous Mie simulation packages are avialble (many of which are listed on
[…]
