### 2026-01-06 #9551 @janakarana
Dear @ayusmin

Thanks for your feedback. Please see my quick response to a few of your comments. Please note that certain details were omitted from the original submission to comply with the 1,000-word limit. Comprehensive documentation, including an overview of core functionality, is available on the project’s wiki page (https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) .

**Comment : On macOS, the application is blocked by Gatekeeper as the developer is not notarized. Either notarization or explicit installation instructions for macOS users are required.**

**Response:** We acknowledge the reviewer's concern regarding macOS Gatekeeper.  To address this without the immediate notarization of the developer account, we have provided explicit, step-by-step installation instructions for macOS users to bypass the Gatekeeper block safely. The Mie Simulator GUI Wiki includes a dedicated "Getting Started on Mac" page (ttps://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Getting-Started-on-Mac) which guides users through the security settings override required for non-notarized applications.  In the manuscript it is mentioned as "Comprehensive documentation, including cross-platform installation guides (Windows, Linux, and macOS), dependency specifications, command-line test execution and several examples, is available on the [Mie Simulator GUI Wiki](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) page.

**Comment: Comparison with existing Mie scattering codes is missing and should be added.**

**Response:** We agree that a direct comparison with established codes is essential. We benchmarked MieSimulatorGUI against Scott Prahl’s Interactive Mie Scattering calculator (https://omlc.org/calc/mie_calc.html), a widely recognized tool in the biomedical optics community. We found perfect numerical agreement. Comparison results for a single wavelength are available on the project wiki page (https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Examples/). In the manuscript it is mentioned as "Comprehensive documentation, including cross-platform installation guides (Windows, Linux, and macOS), dependency specifications, command-line test execution and several examples, is available on the [Mie Simulator GUI Wiki](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) page.
