### 2026-01-08 #9551 @finsberg
@GiannakopoulosIlias and @ayusmin I see that you have generated a second check list. What is the motivation behind that? The JOSS editorial team has developed a new checklist, but this only applies for new submissions (not this one).  We were planning on changing the generated list back to the old one, but we see that you have ticked off all the checkboxes, including some new checkboxes which doesn't apply to this submission (and which I don't think should be ticked off  (for example, the paper does not include an AI usage disclosure). Do you know what has happened here? Thanks again for your time reviewing this submission.
