### 2026-01-23 VirtualPhotonics/MieSimulatorGUI#12 @finsberg
[JOSS REVIEW] Documentation
I think you have a lot of nice documentation. However, there are a few things that I think should be improved.

- [ ] When going to the different installation instructions and it sais download a specific file, then you should add a link to the latest release page
- [ ] Your contributing guide contains information about code style, but not information about how to contribute. Should a new contributor first open an issue and discussing the new contribution? Should a new contributor fork the repository and create a pull request? In the installation instructions you also write about the how to build from source so you could link to this description for developers that want to install a development version of the software (but you should probably ask devlopers to clone the repository rather than downloading the zip file).
- [ ] When referring to other projects (such as QT Creator), e.g when you write _Go to the Qt download page_, then you should provide a link to that page.

This is related to the review at https://github.com/openjournals/joss-reviews/issues/9551
