### 2026-01-23 #9551 @finsberg
Hi @janakarana

Going through the final checklist here.

- I made a small issue about the documentation [here](https://github.com/VirtualPhotonics/MieSimulatorGUI/issues/12). Please address that.
- Create a release of the software and post the version number here. Also archive this release to [Zenodo](https://zenodo.org) (or another place that creates a DOI) and post the DOI here. It is important that the version you archive has the same title and authors as the JOSS paper (also the same LICESE as the repository).
