### 2026-01-24 cd3ed06
Replaced radio buttons in Polar plots by checkboxes; Added units for tooltip data; Added a warning message for dependent scattering; Updated banner banner
