+++ src/app/calc/calculate.cpp
@@ -131,7 +131,7 @@ void Calculate::DoSimulation(QLabel *progress, Parameters *para)
         para->cSca[w]  = sumCsca / sumNumDen ;
         para->cExt[w]  = sumCext / sumNumDen;
         para->cBack[w] = sumCback / sumNumDen;
-        para->SizePara[w] = xPara;
+        para->sizePara[w] = xPara;
         para->mus[w] = sumMus * 1e-6;   //1e-6--> 1micron2 to 1mm2
         para->g[w] = sumMusG /sumMus;
         para->forward[w] = sumForward*100.0/(sumForward+sumBackward);    //Not necessary to divide by sumMus in ratio calculation
@@ -175,7 +175,6 @@ void Calculate::ComputeMuspAtRefWavel(Parameters *para)
     for (unsigned int i = 0; i< 6; i++)
     {
         double lambda = 0.5 + 0.1*i;            // in um
-
 
         sumMus = 0.0;
         sumMusG = 0.0;
@@ -350,30 +349,46 @@ double Calculate::CalculateG(std::complex<double> *S1, std::complex<double> *S2,
 //Set sphere parameters
 void Calculate::SetSphereRadiusAndRefIndex(Parameters *para, unsigned int index, bool flagVolOrConc)
 {
-    double temp, factor;
-    double totalSphereVolume = 0.0;
-    double totalFuncSum = 0.0;
-    double stepR;
-    double *funcArray;
-
     if (para->nRadius == 1)  //Mono Disperse
     {
+        double volFraction;
+        double singleSphereVolume = 4.0 * M_PI *para->meanRadius * para->meanRadius * para->meanRadius / 3.0;
         if (flagVolOrConc)  //If volume fraction is selected, update number density
         {
-           double sphereVolume = 4.0 * M_PI *para->meanRadius * para->meanRadius * para->meanRadius / 3.0;
-           para->sphNumDensity = std::round(1e9 * para->volFraction /sphereVolume) ;
+            para->sphNumDensity = std::round(1e9 * para->volFraction /singleSphereVolume);
+            volFraction = para->volFraction;
         }
+        else
+            volFraction = singleSphereVolume * para->sphNumDensity / 1e9;
+
         para->radArray[0] = para->meanRadius;
         para->numDensityArray[0] = para->sphNumDensity;
         para->scatRefRealArray[0] = para->scatRefReal;
         para->scatRefImagArray[0] = para->scatRefImag;
         para->medRefArray[0] = para->medRef;
+
+        //check independent/dependent scattering (Yalcin ACS Photonics 9(2022))
+        double interParticleDistance = pow(para->sphNumDensity, 1.0/3.0);
+        double clearanceToWavelength = (interParticleDistance - 2 * para->meanRadius)/(1e-3 * para->startWavel);
+        if (volFraction < 0.006 && clearanceToWavelength > 0.5)   //  fv < 0.6% and clearnace/lambda > 0.5
+            para->independentScat = true;
+        else
+            para->independentScat = false;
     }
     else                    //Poly Disperse
     {
-        funcArray = new double[para->nRadius];
+        double temp, factor, stepR;
+        double volFraction;
+        double totalSphereVolume = 0.0;
+        double totalFuncSum = 0.0;
+        double *funcArray = new double[para->nRadius];
+        double *sphereVolume = new double[para->nRadius];
         for (unsigned int i = 0; i < para->nRadius; i++)
-			funcArray[i] = 1.0;
+        {
+            funcArray[i] = 0.0;
+            sphereVolume[i] = 0.0;
+        }
+
         stepR = (para->maxRadius - para->minRadius)/(para->nRadius -1);
 
[…]
