### 2026-01-24 c66a86a
Added new unit test for calculate.cpp
