### 2026-01-24 3b9a459
Added independent scattering condtion check for custom selection; Changed v2.0 pre-release to v2.0
