+++ paper/paper.bib
@@ -295,3 +295,14 @@
 	year = {1980},
 	doi = {10.1364/AO.19.001505},
 }
+@article{Yalcin2022,
+	title={Dependent Scattering in Thick and Concentrated Colloidal Suspensions},
+	author={Yalcin, R. A. and Lee, T. and Kashanchi, G. Na. and Markkanen, J. and Martinez, R. and Tolbert, S. H. and Pilon, L.},
+	journal={ACS Photonics},
+	volume={9},
+	number={10},
+	pages={3318--3332},
+	year={2022},
+	publisher={ACS Publications},
+	doi={10.1021/acsphotonics.2c00664},
+}
