+++ paper/paper.md
@@ -67,7 +67,7 @@ The tool is distributed as portable binaries for Windows, macOS, and Linux. For 
 
 ## Input Selection Panel
 
-This panel enables the user to define inputs for Mie simulations at either a single wavelength or across a spectral range. The distribution of spheres is described using sphere concentration (`Conc`) ($\text{spheres/mm}^3$) or volume fraction (`Vol Frac`). The volume fraction, the proportion of space occupied by spheres, is determined by the product of the sphere concentration in 1 $\text{mm}^3$ and the individual sphere volume. `MieSimulatorGUI` utilizes the independent scattering approximation, a framework valid for dilute suspensions where particles are sufficiently separated to ignore coherent interactions [@VandeHulst1957; @Schmitt1998]. Since the validity of this approximation depends on the inter-particle spacing, size parameter, and relative refractive index, its accuracy diminishes as the volume fraction increases [@Ivezic1996; @Galy2020]. Consequently, the tool is best suited for dilute systems; the results obtained for concentrated regimes may deviate from physical reality and should be interpreted with caution.
+This panel enables the user to define inputs for Mie simulations at either a single wavelength or across a spectral range. The distribution of spheres is described using sphere concentration (`Conc`) ($\text{spheres/mm}^3$) or volume fraction (`Vol Frac`). The volume fraction, the proportion of space occupied by spheres, is determined by the product of the sphere concentration in 1 $\text{mm}^3$ and the individual sphere volume. `MieSimulatorGUI` utilizes the independent scattering approximation, a framework valid for dilute suspensions where particles are sufficiently separated to ignore coherent interactions [@VandeHulst1957; @Schmitt1998]. Since the validity of this approximation depends on the inter-particle spacing, size parameter, and relative refractive index, its accuracy diminishes as the volume fraction increases [@Ivezic1996; @Galy2020; Yalcin2022]. Consequently, the tool is best suited for dilute systems; the results obtained for concentrated regimes may deviate from physical reality and should be interpreted with caution.
 
 To maintain numerical stability in the BHMIE algorithm and ensure UI responsiveness, sphere diameters are restricted to a range of 0.1 $\text{nm}$ to 300 $\text{µm}$ and wavelengths are limited to $50\text{ nm}$ – $3000\text{ nm}$. These ranges cover the primary biomedical and atmospheric spectral windows. For absorbing spheres, the complex refractive index ($m_{sphere}$) is defined as $m_{real}$ – j $m_{imag}$, where $m_{real}$ and $m_{imag}$ represent real and imaginary components, respectively [@VandeHulst1957; @Wiscombe1979]. 
 
