### 2026-01-24 2469164
Added more tests for functions in calculate.cpp; Fixed display scaling issue
