+++ .github/workflows/test.yml
@@ -37,12 +37,3 @@ jobs:
         working-directory: ./src/build/test
         run: chmod +x MieSimulatorGUI_test
              ./MieSimulatorGUI_test
-
-      - name: Delete successful runs
-        uses: Mattraks/delete-workflow-runs@v2
-        with:
-            token: ${{ github.token }}
-            repository: ${{ github.repository }}
-            retain_days: 0
-            keep_minimum_runs: 0
-            status: success  # Only deletes successful runs
