+++ CONTRIBUTING.md
@@ -1,79 +1,117 @@
-When contributing to the Mie Simulator GUI please follow the coding conventions detailed below. Some of the initial development did not adhere to these guidelines but going forward, this will be the code style and structure.
-## Naming Conventions
+# Contribution Workflow
+
+All contributions to the **MieSimulatorGUI** must follow the workflow outlined below to ensure changes are traceable, reviewable, and consistent.
+
+## 1. Create an Issue First
+Before starting any work:
+* **Create a GitHub issue** describing the bug, enhancement, or feature request.
+* Clearly explain the motivation, scope, and expected behavior.
+* Wait for confirmation or discussion if the change is significant.
+* *Every code change must be associated with an issue.*
+
+## 2. Fork the Repository
+* Fork the main repository into your own GitHub account.
+* All development work must be done in your fork, not directly in the main repository.
+
+## 3. Create a Feature Branch
+Create a new branch from the default branch (e.g., `master`). The branch name must be derived from the issue and prepended with `feature/`.
+
+**Branch naming format:**
+`feature/<issue-number>-short-description`
+
+**Example:**
+`feature/42-add-particle-size-slider`
+
+## 4. Implement the Change
+* Follow all coding conventions defined in this document.
+* Keep changes focused on the scope of the issue.
+* Ensure the code builds and tests pass before submitting.
+
+## 5. Open a Pull Request
+Open a Pull Request (PR) from your feature branch to the main repository. Reference the related issue in the PR description (e.g., `Closes #42`).
+
+**Clearly describe:**
+* What was changed
+* Why the change was made
+* Any relevant design or implementation details
+
+## 6. Code Review and Approval
+* All pull requests require review before merging.
+* Be prepared to address feedback or requested changes.
+* Once approved, the pull request will be merged by a repository maintainer.
+
+> [!IMPORTANT]
+> Direct commits to the main repository are not permitted.
+
+# Coding Convention
+When contributing to the **MieSimulatorGUI**, please follow the coding conventions detailed below. These guidelines represent the established style observed in the core simulation engine (e.g., `Calculate` and `MieSimulation` classes).
+
+## 1. Naming Conventions
+
 ### Definitions:
-* __Camel Case:__ Naming convention where the first letter of each word is capitalized with the exception of the first letter. Example: __thisIsCamelCased__
-* __Pascal Case:__ Naming convention where the first letter of each word is capitalized including the first letter. Example: __ThisIsPascalCased__
-* __Underscore Uppercase Delimited:__ Naming convention that places underscores between the words and the first letter of each word is capitalized. Example: __This_Is_Underscore_Delimited__
-* __Underscore Lowercase Delimited:__ Naming convention that places underscores between the words and the first letter of each word is not capitalized. Example: __this_is_underscore_lowercase_delimited__<br />
+* **Camel Case:** Naming convention where the first letter of each word is capitalized with the exception of the first letter. Example: `thisIsCamelCase`
+* **Pascal Case:** Naming convention where the first letter of every word is capitalized. Example: `ThisIsPascalCase`
+* **M-Prefix Pascal Case:** Member variables prefixed with a lowercase 'm' followed by the variable name in Pascal Case. Example: `mMemberVariable`
 
-These are the only naming conventions that should be used in the Mie Simulator GUI and apply to different aspects of the code defined below.
-* __Classes__, __properties__, and __functions__, should be Pascal cased. Example: __MyClass__
[…]
