### 2026-01-25 84beda7
Fixed Qt version number; Corrected colors for dark theme windows; Fixed Tooltip display error; Defined different fonts for different OS
