+++ src/app/dialog/plotdata.cpp
@@ -523,11 +523,11 @@ void PlotData::PlotPhaseFunctionLinear(Ui_MainWindow *ui, QVector<double> x, QVe
         }
         if (ui->checkBox_PhaseLinearPara->isChecked())
         {
-            PlotSingleGraph(customPlot, x, yPara, Qt::blue, "Para.", idx, 2);
+            PlotSingleGraph(customPlot, x, yPara, QColor(0, 135, 255), "Para.", idx, 2);
             idx++;
         }
         if (ui->checkBox_PhaseLinearPerp->isChecked())
-            PlotSingleGraph(customPlot, x, yPerp, Qt::darkGreen, "Perp.", idx, 2);
+            PlotSingleGraph(customPlot, x, yPerp, QColor(0, 150, 0), "Perp.", idx, 2);
 
         if (ui->radioButton_LinearYAxis->isChecked())
             customPlot->yAxis->setLabel("Magnitude");
@@ -583,9 +583,9 @@ void PlotData::PlotPhaseFunctionPolar(Ui_MainWindow *ui, QVector<double> theta, 
         if (ui->checkBox_PhasePolarAve->isChecked())
             PlotSingleCurve(customPlot, x1, y1, Qt::red, "Ave.", totalSize);
         if (ui->checkBox_PhasePolarPara->isChecked())
-            PlotSingleCurve(customPlot, x2, y2, Qt::blue, "Para.", totalSize);
+            PlotSingleCurve(customPlot, x2, y2, QColor(0, 135, 255), "Para.", totalSize);
         if (ui->checkBox_PhasePolarPerp->isChecked())
-            PlotSingleCurve(customPlot, x3, y3, Qt::darkGreen, "Perp.", totalSize);
+            PlotSingleCurve(customPlot, x3, y3, QColor(0, 150, 0), "Perp.", totalSize);
 
         // // Legend configuration
         // customPlot->legend->setWrap(3);
@@ -627,7 +627,7 @@ void PlotData::PlotS1S2(Ui_MainWindow *ui, Parameters *para, QVector<double> x, 
     if (ui->radioButton_S1S2->isChecked())
     {
         PlotSingleGraph(customPlot, x, yS1, Qt::red, "S1", 0, 2);
-        PlotSingleGraph(customPlot, x, yS2, Qt::blue, "S2", 1, 2);
+        PlotSingleGraph(customPlot, x, yS2, QColor(0, 135, 255), "S2", 1, 2);
         customPlot->yAxis->setLabel("S1,S2");
         if (ui->radioButton_LogYAxis->isChecked())
             customPlot->yAxis->setLabel("Log (S1,S2)");
@@ -680,7 +680,7 @@ void PlotData::PlotMuspPowerLaw(Ui_MainWindow *ui, QVector<double> x, QVector<do
     PlotSingleGraph(customPlot, x, yMusp, Qt::red, "μs' Data", 0, 4);
 
     //Power law fit plot
-    PlotSingleGraph(customPlot, x, yFit, Qt::blue, "Best Fit", 1, 4);
+    PlotSingleGraph(customPlot, x, yFit, QColor(0, 135, 255), "Best Fit", 1, 4);
 
     customPlot->legend->setBrush(QBrush(QColor(255, 255, 255, 0)));
     customPlot->legend->setBorderPen(QPen(Qt::NoPen));
@@ -834,7 +834,7 @@ void PlotData::PlotForwardBackward(Ui_MainWindow *ui, QVector<double> x, QVector
 
     //Forward and backward Plot
     PlotSingleGraph(customPlot, x, yF, Qt::red, "Forward Scat. %", 0, 4);
[…]
