### 2026-01-25 4bb702a
Cleaned up code
