+++ src/app/dialog/optionsdialog.cpp
@@ -10,11 +10,11 @@
 OptionsDialog::OptionsDialog(QWidget *parent) :
     QDialog(parent),
     ui(new Ui::OptionsDialog)
-{    
-    flagScatPara = false;
-    flagPhaseFunction = false;
-    flagS1 = false;
-    flagS2 = false;
+{
+    mFlagScatPara = false;
+    mFlagPhaseFunction = false;
+    mFlagS1 = false;
+    mFlagS2 = false;
     ui->setupUi(this);
 }
 
@@ -25,25 +25,25 @@ OptionsDialog::~OptionsDialog()
 
 void OptionsDialog::on_pushButton_ScatPara_clicked()
 {
-    flagScatPara = true;
+    mFlagScatPara = true;
     this->close();
 }
 
 void OptionsDialog::on_pushButton_PhaseFunction_clicked()
 {
-    flagPhaseFunction = true;
+    mFlagPhaseFunction = true;
     this->close();
 }
 
 void OptionsDialog::on_pushButton_S1_clicked()
 {
-    flagS1 = true;
+    mFlagS1 = true;
     this->close();
 }
 
 void OptionsDialog::on_pushButton_S2_clicked()
 {
-    flagS2 = true;
+    mFlagS2 = true;
     this->close();
 }
 
@@ -67,14 +67,16 @@ void OptionsDialog::SaveData(QRadioButton *radioButton_MonoDisperse,
     exec();
 
     double margin = (1.0 + slider_ConcPercentChange->value() /200.0);
-    if (flagScatPara)
+    if (mFlagScatPara)
     {
         QString fileName = QFileDialog::getSaveFileName(this,
                  tr("Save Data"), "Mie_ScatteringParameters",
                  tr("Text File (*.txt);;All Files (*)"));
 
         if (fileName.isEmpty())
+        {
             return;
+        }
         else
         {
             SaveScatPara(radioButton_MonoDisperse,
@@ -86,14 +88,16 @@ void OptionsDialog::SaveData(QRadioButton *radioButton_MonoDisperse,
 
 
     }
-    if (flagPhaseFunction)
+    if (mFlagPhaseFunction)
     {
         QString fileName = QFileDialog::getSaveFileName(this,
                  tr("Save Data"), "Mie_PhaseFunctionData",
                  tr("Text File (*.txt);;All Files (*)"));
 
         if (fileName.isEmpty())
+        {
             return;
+        }
         else
         {
             SavePhaseFunction(checkBox_PhasePolarAve,
@@ -103,28 +107,32 @@ void OptionsDialog::SaveData(QRadioButton *radioButton_MonoDisperse,
             RememberLastDirectory(fileName);
         }
     }
-    if (flagS1)
+    if (mFlagS1)
     {
         QString fileName = QFileDialog::getSaveFileName(this,
                  tr("Save Data"), "Mie_S1Data",
                  tr("Text File (*.txt);;All Files (*)"));
 
         if (fileName.isEmpty())
+        {
             return;
+        }
         else
         {
             SaveS1(para,fileName);
             RememberLastDirectory(fileName);
         }
     }
-    if (flagS2)
+    if (mFlagS2)
     {
         QString fileName = QFileDialog::getSaveFileName(this,
                  tr("Save Data"), "Mie_S2Data",
                  tr("Text File (*.txt);;All Files (*)"));
 
         if (fileName.isEmpty())
+        {
             return;
+        }
         else
         {
             SaveS2(para,fileName);
@@ -161,44 +169,61 @@ void OptionsDialog::SaveScatPara(QRadioButton *radioButton_MonoDisperse,
     {
[…]
