+++ src/app/dialog/mainwindow.cpp
@@ -258,11 +258,11 @@ void MainWindow::on_pushButton_ShowDistributionAndCustom_clicked()
                     totalNumDensity += mPara->numDensityArray[i];
                     totalRadius += mPara->radArray[i];
                 }
-                double volFraction = totalSphereVolume/1e9;
-                double meanRadius = totalRadius/mPara->nRadius;
+                double volFraction = totalSphereVolume / 1e9;
+                double meanRadius = totalRadius / mPara->nRadius;
 
                 //check independent/dependent scattering (Yalcin ACS Photonics 9(2022))
-                double interParticleDistance = pow(totalNumDensity, 1.0/3.0);
+                double interParticleDistance = pow(totalNumDensity, 1.0 / 3.0);
                 double clearanceToWavelength = (interParticleDistance - 2 * meanRadius)/(1e-3 * mPara->startWavel);
                 if (volFraction < 0.006 && clearanceToWavelength > 0.5)   //  fv < 0.6% and clearnace/lambda > 0.5
                     mPara->independentScat = true;
@@ -783,7 +783,7 @@ void MainWindow::on_checkBox_PhasePolarPerp_toggled()
 void MainWindow::on_slider_ConcPercentChange_valueChanged(int position)
 {
     double value;
-    double margin = (1.0 + position /200.0);
+    double margin = (1.0 + position / 200.0);
     ui->label_ActualConcPercent->setText(QString::number(position/2.0)+"%");
 
     if (ui->radioButton_NumDen->isChecked())
@@ -799,7 +799,10 @@ void MainWindow::on_slider_ConcPercentChange_valueChanged(int position)
 
     PlotData plot;
     if (mDistPlotFlag)
+    {
         plot.AssignValuesDistributionPlot(ui, mPara);
+    }
+
     if (mOtherPlotsFlag)
     {   MainWindowSupport support;
         support.DisableEnableRealImagButtons(ui);
@@ -812,6 +815,7 @@ void MainWindow::on_slider_WL_PFPolar_valueChanged(int value)
 {
     int wavel = static_cast<int>(mPara->startWavel + value*mPara->stepWavel);
     ui->label_CurrentWL_PFPolar->setText(QString::number(wavel));
+
     if (mOtherPlotsFlag)
     {
         PlotData plot;
@@ -848,6 +852,7 @@ void MainWindow::on_comboBox_Distribution_currentIndexChanged(int value)
     // 0: Log Normal
     // 1: Gaussian
     // 2: Custom
+
     if (value == mPara->Custom) //Custom Distribution
     {
         support.DisableWidgetsDuringCustomPolyDisperseData(ui, true);
@@ -870,10 +875,15 @@ void MainWindow::MouseOverPlotScatteringCrossSection(QMouseEvent *event)
     QString strNameY;
     QString strUnitX = " nm";
     QString strUnitY = " \u00B5m\u00B2";
+
     if (ui->radioButton_LogYAxis->isChecked())
+    {
         strNameY = "Log(Csca)";
+    }
     else
+    {
         strNameY = "Csca";
+    }
     DisplayGraphData(event, customPlot, strNameX, strUnitX, strNameY, strUnitY);
 }
 
@@ -886,10 +896,15 @@ void MainWindow::MouseOverPlotExtinctionCrossSection(QMouseEvent *event)
     QString strNameY;
     QString strUnitX = " nm";
     QString strUnitY = " \u00B5m\u00B2";
+
     if (ui->radioButton_LogYAxis->isChecked())
+    {
         strNameY = "Log(Cext)";
+    }
     else
+    {
         strNameY = "Cext";
+    }
[…]
