### 2026-01-27 a4aa0b2
Added a function to check the packing volume for polydisperse distribution
