### 2026-01-28 3b635b7
Corrected independent/dependent scattering logic and included tests
