+++ src/app/dialog/mainwindow.cpp
@@ -161,18 +161,13 @@ void MainWindow::on_pushButton_RunSimulation_clicked()
             //Mono disperse
             if (ui->radioButton_MonoDisperse->isChecked())
             {
-                support.ProcessDistribution(ui, mPara, mPara->MonoDisperse);  // index = 3: mono disperse
-
+                support.ProcessDistribution(ui, mPara, mPara->MonoDisperse);  // index = 3: monodisperse
                 support.ProcessMonoDisperse(ui,mPara);
                 ui->label_Progress->setText("<font color=\"green\"> Completed!</font>");
                 ui->slider_ConcPercentChange->setValue(0);
                 mOtherPlotsFlag = true;
                 mDistPlotFlag = true;
-                if(!mPara->independentScat)
-                    DisplayWarning("The concentration or volume fraction exceeds the "
-                                   "limits for the independent scattering approximation. "
-                                   "As this tool is best suited for dilute systems, "
-                                   "results should be interpreted with caution.");
+
                 //Enable widgets
                 support.DisableWidgetsDuringSimulation(ui, mPara, false);
             }
@@ -182,17 +177,15 @@ void MainWindow::on_pushButton_RunSimulation_clicked()
             {
                 if (mPara->CheckDistributionParameters(ui->comboBox_Distribution))   //sanity check
                 {
-                    support.ProcessDistribution(ui, mPara, static_cast<unsigned int>(ui->comboBox_Distribution->currentIndex()));                    
-                    support.ProcessPolyDisperse(ui,mPara);
-                    ui->label_Progress->setText("<font color=\"green\"> Completed!</font>");
-                    ui->slider_ConcPercentChange->setValue(0);
-                    mOtherPlotsFlag = true;
-                    mDistPlotFlag = true;
-                    if(!mPara->independentScat)
-                        DisplayWarning("The concentration or volume fraction exceeds the "
-                                       "limits for the independent scattering approximation. "
-                                       "As this tool is best suited for dilute systems, "
-                                       "results should be interpreted with caution.");
+                    support.ProcessDistribution(ui, mPara, static_cast<unsigned int>(ui->comboBox_Distribution->currentIndex()));
+                    if (mPara->CheckPackingVolume())   // check packing volume for polydisperse
+                    {
+                        support.ProcessPolyDisperse(ui,mPara);
+                        ui->label_Progress->setText("<font color=\"green\"> Completed!</font>");
+                        ui->slider_ConcPercentChange->setValue(0);
+                        mOtherPlotsFlag = true;
+                        mDistPlotFlag = true;
+                    }
                 }
                 //Enable widgets
                 support.DisableWidgetsDuringSimulation(ui, mPara, false);
@@ -245,29 +238,6 @@ void MainWindow::on_pushButton_ShowDistributionAndCustom_clicked()
                 support.SetWidgets(ui, mPara);
                 support.ProcessDistribution(ui, mPara, static_cast<unsigned int>(ui->comboBox_Distribution->currentIndex()));
                 mDistPlotFlag = true;
-
-                // compute parameters required to check dependent/independent scattering
-                double totalSphereVolume = 0.0;
-                double totalNumDensity = 0.0;
-                double totalRadius = 0.0;
-                for (unsigned int i=0; i<mPara->nRadius; i++)
-                {
-                    double singleSphereVolume = 4.0 * M_PI * mPara->radArray[i] *
-                                      mPara->radArray[i] * mPara->radArray[i] /3.0;
[…]
