### 2026-01-29 20883fd
Included a new reference and added a sentence regarding the dependent scattering warning message.
