+++ paper/paper.bib
@@ -251,6 +251,15 @@
 	pages = {3161--3187},
 	doi = {10.5194/amt-15-3161-2022}
 }
+@article{Tien1987,
+	author  = {Tien, C. L. and Drolen, B. L.},
+	title   = {Thermal radiation in particulate media with dependent and independent scattering},
+	journal = {Annual Review of Numerical Fluid Mechanics and Heat Transfer},
+	year    = {1987},
+	volume  = {1},
+	pages   = {1--32},
+	doi     = {10.1615/AnnualRevHeatTransfer.v1.30}
+}
 @book{VandeHulst1957,
 	title        = {{Light scattering by small particles}},
 	author       = {van de Hulst, H. C.},
