+++ paper/paper.md
@@ -54,7 +54,7 @@ While numerous Mie simulation packages are available (many of which are listed o
 
 # Main Features
 
-Built on the BHMIE [@Bohren1983] and Wiscombe [@Wiscombe1979] frameworks, `MieSimulatorGUI` calculates spectral optical properties, including scattering coefficients, cross-sections, scattering amplitude matrix entries ($\text{S}_1$, $\text{S}_2$), phase functions, and scattering asymmetry. The tool supports both monodisperse and polydisperse particle distributions [@Gelebart1996] and facilitates parameter estimation by fitting reduced scattering coefficient curves, a technique of significant value in tissue optics [@Jacques2013]. To ensure computational accuracy and GUI stability, the software includes an automated test suite integrated via GitHub Actions. Its high-performance C++ engine enables near-instantaneous computation and plotting, providing real-time visual updates across the spectral range. Comprehensive documentation, including cross-platform installation guides (Windows, Linux, and macOS), dependency specifications, command-line test execution and several examples, is available on the [Mie Simulator GUI Wiki](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) page. 
+Built on the BHMIE [@Bohren1983] and Wiscombe [@Wiscombe1979] frameworks, `MieSimulatorGUI` calculates spectral optical properties, including scattering coefficients, cross-sections, scattering amplitude matrix entries ($\text{S}_1$, $\text{S}_2$), phase functions, and scattering asymmetry. The tool supports both monodisperse and polydisperse particle distributions [@Gelebart1996] and facilitates parameter estimation by fitting reduced scattering coefficient curves, a technique of significant value in tissue optics [@Jacques2013]. To ensure computational accuracy and GUI stability, the software includes an automated test suite integrated via GitHub Actions. Its high-performance C++ engine enables near-instantaneous computation and plotting, providing real-time visual updates across the spectral range. Comprehensive documentation, including cross-platform installation guides (Windows, Linux, and macOS), dependency specifications, command-line test execution, dependent scattering warning trigger conditions, and several examples, is available on the [Mie Simulator GUI Wiki](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki) page. 
 
 ![Figure 1: Six interactive panels of `MieSimulatorGUI`: (a) Input selection, (b) Particle size distribution, (c) Scattering coefficient, (d) Phase function, (e) Reduced Scattering,  and (f) Scattering Asymmetry\label{fig:Figure1}](Figure1.png){ width=100% }
 Figure 1: Six interactive panels of `MieSimulatorGUI`: (a) Input selection, (b) Particle size distribution, (c) Scattering coefficient, (d) Phase function, (e) Reduced Scattering, and (f) Scattering Asymmetry (Anisotropy)
@@ -67,7 +67,7 @@ The tool is distributed as portable binaries for Windows, macOS, and Linux. For 
 
 ## Input Selection Panel
 
[…]
