### 2026-01-29 VirtualPhotonics/MieSimulatorGUI#12 @janakarana
1. We have updated all installation instructions across the documentation to ensure that any mention of a specific file download is accompanied by a direct link to the Latest Release page.
2. We have significantly expanded our CONTRIBUTING.md file to move beyond code style. It now outlines a clear operational workflow. We have created a dedicated [Contribution Workflow Wiki Page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Contribution-Workflow) to guide new developers through the process.
3. To improve navigability, all references to external tools and projects now include direct hyperlinks.
