### 2026-01-30 025918a
Corrected dependent display logic and warning message; Fixed polydisperse distribution
