+++ src/app/calc/calculate.h
@@ -31,7 +31,9 @@ public:
     double CalculateForwardBackward(std::complex<double> *S1, std::complex<double> *S2,
                                     Parameters *para, unsigned int start, unsigned int end);
     double CalculateG(std::complex<double> *S1, std::complex<double> *S2, Parameters *para);
-    bool CheckIndependentScattering(Parameters *para);
+    bool CheckIndependentScattering(Parameters *para, double &clearanceToWavelength,
+                                    double &sizeParameter, double &volFraction,
+                                    double &criticalWavength, QString &strRegime);
 };
 
 #endif // CALCULATE_H

[…]
