### 2026-01-30 6573aea
Increased number of sphere limit from 101 to 201
