+++ src/app/calc/calculate.h
@@ -33,7 +33,7 @@ public:
     double CalculateG(std::complex<double> *S1, std::complex<double> *S2, Parameters *para);
     bool CheckIndependentScattering(Parameters *para, double &clearanceToWavelength,
                                     double &sizeParameter, double &volFraction,
-                                    double &criticalWavength, QString &strRegime);
+                                    double &criticalWavelength, QString &strRegime);
 };
 
 #endif // CALCULATE_H
