### 2026-01-31 92ea63d
Color coded the dependent scattering warning message
