+++ paper/paper.md
@@ -67,7 +67,7 @@ The tool is distributed as portable binaries for Windows, macOS, and Linux. For 
 
 ## Input Selection Panel
 
-This panel enables the user to define inputs for Mie simulations at either a single wavelength or across a spectral range. The distribution of spheres is described using sphere concentration (`Conc`) ($\text{spheres/mm}^3$) or volume fraction (`Vol Frac`). The volume fraction represents the ratio of total sphere volume to the total container volume. It is calculated by multiplying the volume of an individual sphere by `Conc`.`MieSimulatorGUI` utilizes the independent scattering approximation, a framework valid for dilute suspensions where particles are sufficiently separated to ignore coherent interactions [@VandeHulst1957; @Schmitt1998]. Since the validity of this approximation depends on the inter-particle spacing, size parameter, and relative refractive index [@Tien1987; @Ivezic1996; @Galy2020; @Yalcin2022], its accuracy diminishes as the volume fraction increases. Consequently, the tool is best suited for dilute systems and the results obtained for concentrated regimes may deviate from physical reality and should be interpreted with caution. `MieSimulatorGUI` triggers a warning message whenever input parameters shift into the transitional or dependent scattering regimes.
+This panel enables the user to define inputs for Mie simulations at either a single wavelength or across a spectral range. The distribution of spheres is described using sphere concentration (`Conc`) ($\text{spheres/mm}^3$) or volume fraction (`Vol Frac`). The volume fraction represents the ratio of total sphere volume to the total container volume. For polydisperse systems, it is calculated by summing the products of each sphere size's volume and its respective concentration per unit volume. `MieSimulatorGUI` utilizes the independent scattering approximation, a framework valid for dilute suspensions where particles are sufficiently separated to ignore coherent interactions [@VandeHulst1957; @Schmitt1998]. Since the validity of this approximation depends on the inter-particle spacing, size parameter, and relative refractive index [@Tien1987; @Ivezic1996; @Galy2020; @Yalcin2022], its accuracy diminishes as the volume fraction increases. Consequently, the tool is best suited for dilute systems and the results obtained for concentrated regimes may deviate from physical reality and should be interpreted with caution. `MieSimulatorGUI` triggers a warning when parameters violate dependent scattering conditions.
 
 To maintain numerical stability in the BHMIE algorithm and ensure UI responsiveness, sphere diameters are restricted to a range of 0.1 $\text{nm}$ to 300 $\text{µm}$ and wavelengths are limited to $50\text{ nm}$ – $3000\text{ nm}$. These ranges cover the primary biomedical and atmospheric spectral windows. For absorbing spheres, the complex refractive index ($m_{sphere}$) is defined as $m_{real}$ – j $m_{imag}$, where $m_{real}$ and $m_{imag}$ represent real and imaginary components, respectively [@VandeHulst1957; @Wiscombe1979]. 
 
@@ -84,19 +84,19 @@ For monodisperse distribution, the [scattering coefficient](https://omlc.org/cla
 
 ## Phase Function Panel
 
[…]
