### 2026-01-31 ce12ee4
Set window size to 1120x768
