### 2026-01-31 400ede7
Fixed incorrect VolFraction usage
