+++ paper/paper.bib
@@ -120,6 +120,19 @@
 	issn         = {0031-9155},
 	url          = {http://stacks.iop.org/0031-9155/58/i=14/a=5007?key=crossref.c531ac37cfa1f77bbc0d33a8e92de8c9}
 }
+@article{Kodach2011,
+	author = {V. M. Kodach and D. J. Faber and J. van Marle and T. G. van Leeuwen and J. Kalkman},
+	journal = {Opt. Express},	
+	number = {7},
+	pages = {6131--6140},
+	publisher = {Optica Publishing Group},
+	title = {Determination of the scattering anisotropy with optical coherence tomography},
+	volume = {19},
+	month = {Mar},
+	year = {2011},
+	url = {https://opg.optica.org/oe/abstract.cfm?URI=oe-19-7-6131},
+	doi = {10.1364/OE.19.006131},
+}
 @article{Majic2020,
 	title        = {{Numerically stable formulation of Mie theory for an emitter close to a sphere}},
 	author       = {Majic, Matt and {Le Ru}, Eric C.},
@@ -267,6 +280,16 @@
 	publisher    = {John Wiley and Sons Inc.},
 	address      = {New York},
 	pages        = 470
+}
+@article{vanStaveren1991,
+	author  = {van Staveren, Hugo J. and Moes, Christian J. M. and van Marie, Jan and Prahl, Scott A. and van Gemert, Martin J. C.},
+	title   = {Light scattering in lntralipid-10\% in the wavelength range of 400–1100 nm},
+	journal = {Applied Optics},
+	year    = {1991},
+	volume  = {30},
+	number  = {31},
+	pages   = {4507--4514},
+	doi     = {10.1364/ao.30.004507}
 }
 @article{Wang2005,
 	title        = {{Approximation of Mie scattering parameters in near-infrared tomography of normal breast tissue in vivo}},
