+++ paper/paper.md
@@ -15,7 +15,7 @@ authors:
   - name: Carole K. Hayakawa
     orcid: 0000-0001-5696-160X
     affiliation: '1, 2'
-  - name: Lisa Glover
+  - name: Lisa M. Glover
     orcid: 0009-0009-0291-5099
     affiliation: 2
   - name: Vasan Venugopalan
@@ -32,7 +32,7 @@ affiliations:
       Beckman Laser Institute and Medical Clinic, University of California,
       Irvine, California 92697, United States
     index: 2
-date: 11 December 2025
+date: 2 February 2026
 bibliography: paper.bib
 
 ---
@@ -67,7 +67,7 @@ The tool is distributed as portable binaries for Windows, macOS, and Linux. For 
 
 ## Input Selection Panel
 
-This panel enables the user to define inputs for Mie simulations at either a single wavelength or across a spectral range. The distribution of spheres is described using sphere concentration (`Conc`) ($\text{spheres/mm}^3$) or volume fraction (`Vol Frac`). The volume fraction represents the ratio of total sphere volume to the total container volume. For polydisperse systems, it is calculated by summing the products of each sphere size's volume and its respective concentration per unit volume. `MieSimulatorGUI` utilizes the independent scattering approximation, a framework valid for dilute suspensions where particles are sufficiently separated to ignore coherent interactions [@VandeHulst1957; @Schmitt1998]. Since the validity of this approximation depends on the inter-particle spacing, size parameter, and relative refractive index [@Tien1987; @Ivezic1996; @Galy2020; @Yalcin2022], its accuracy diminishes as the volume fraction increases. Consequently, the tool is best suited for dilute systems and the results obtained for concentrated regimes may deviate from physical reality and should be interpreted with caution. `MieSimulatorGUI` triggers a warning when parameters violate dependent scattering conditions.
+This panel enables the user to define inputs for Mie simulations at either a single wavelength or across a spectral range. The distribution of spheres is described using sphere concentration (`Conc`) ($\text{spheres/mm}^3$) or volume fraction (`Vol Frac`). The volume fraction represents the ratio of total sphere volume to the total container volume. For polydisperse systems, it is calculated by summing the products of each sphere size's volume and its respective concentration per unit volume. `MieSimulatorGUI` utilizes the independent scattering approximation, a framework valid for dilute suspensions where particles are sufficiently separated to ignore coherent interactions [@VandeHulst1957; @Schmitt1998]. Since the validity of this approximation depends on the inter-particle spacing, size parameter, and relative refractive index [@Tien1987; @Ivezic1996; @Galy2020; @Yalcin2022], its accuracy diminishes as the volume fraction increases. Consequently, the tool is best suited for dilute systems and the results obtained for concentrated regimes may deviate from physical reality and should be interpreted with caution. `MieSimulatorGUI` v2.0 triggers a warning when parameters violate dependent scattering conditions.
 
 To maintain numerical stability in the BHMIE algorithm and ensure UI responsiveness, sphere diameters are restricted to a range of 0.1 $\text{nm}$ to 300 $\text{µm}$ and wavelengths are limited to $50\text{ nm}$ – $3000\text{ nm}$. These ranges cover the primary biomedical and atmospheric spectral windows. For absorbing spheres, the complex refractive index ($m_{sphere}$) is defined as $m_{real}$ – j $m_{imag}$, where $m_{real}$ and $m_{imag}$ represent real and imaginary components, respectively [@VandeHulst1957; @Wiscombe1979]. 
 
@@ -96,10 +96,10 @@ This panel shows the [reduced scattering coefficient](https://omlc.org/classroom
 
[…]
