+++ paper/paper.md
@@ -96,7 +96,7 @@ This panel shows the [reduced scattering coefficient](https://omlc.org/classroom
 
 ## Example Application: Scattering of Intralipid Phantoms
 
-To demonstrate the tool's scientific utility, we considered the characterization of Intralipid, a standard tissue phantom in biomedical optics [@vanStaveren1991; @DiNinni2011]. Based on Intralipid particle distribution profiles in the literature [@Kodach2011; @Raju2017], we assumed a polydisperse `Log Normal` particle distribution with a mean diameter of $0.22 \mu\text{m}$ and a standard deviation of $0.37 \mu\text{m}$. We set the refractive indices to 1.47 for the soybean oil and 1.33 for the surrounding medium, while assigning a value of 101 to the `Num. sph. sizes` field. To analyze different concentrations ranging from 0.2% to 20% [@Aernouts2013; @vanStaveren1991], volume fractions were scaled using a baseline value of 0.227 for a 20% (w/w) Intralipid concentration [@Aernouts2013]. Upon executing the simulation across the 400–2250 $\text{nm}$ spectral range, `MieSimulatorGUI` calculates $\text{µ}_s$, $\text{µ}_s'$ and g. While the selected volume fractions may exceed independent scattering limits  in the literature [@Tien1987; @Galy2020; @Yalcin2022], the results remain comparable with previously reported bulk scattering properties [@vanStaveren1991; @Aernouts2013] of Intralipid. Alternatively, users may utilize the `Custom` option to upload specific sphere profiles such as those from Raju et al. [@Raju2017], and export the resulting data as a text file for further analysis.
+To demonstrate the tool's scientific utility, we considered the characterization of Intralipid, a standard tissue phantom in biomedical optics [@vanStaveren1991; @DiNinni2011]. Based on Intralipid particle distribution profiles in the literature [@Kodach2011; @Raju2017], we assumed a polydisperse `Log Normal` particle distribution with a mean diameter of $0.22 \mu\text{m}$ and a standard deviation of $0.37 \mu\text{m}$. We set the refractive indices to 1.47 for the soybean oil droplets and 1.33 for the surrounding medium, while assigning a value of 101 to the `Num. sph. sizes` field. To analyze different concentrations ranging from 0.2% to 20% [@Aernouts2013; @vanStaveren1991], volume fractions were scaled using a baseline value of 0.227 for a 20% (w/w) Intralipid concentration [@Aernouts2013]. Upon executing the simulation across the 400–2250 $\text{nm}$ spectral range, `MieSimulatorGUI` calculates $\text{µ}_s$, $\text{µ}_s'$ and g. While the selected volume fractions may exceed independent scattering limits established in the literature [@Tien1987; @Galy2020; @Yalcin2022], the results show strong agreement with established bulk scattering properties [@vanStaveren1991; @Aernouts2013]. Figures can be exported as .png files and results as text files for further analysis. Beyond preset distributions, the `Custom` option can be utilized to upload specific sphere profiles [@Raju2017].
 
 # Acknowledgments
 
