### 2026-02-04 fb2add3
Updated dependent scattering warning message
