+++ src/app/dialog/mainwindowsupport.h
@@ -23,11 +23,10 @@ public:
     void DisableWidgetsDuringSimulation(Ui_MainWindow *ui, Parameters *para, bool flag);
     void DisableWidgetsDuringCustomPolyDisperseData(Ui_MainWindow *ui, bool flag);
     void ReadCustomData(Parameters * para, QString fileName, bool * dataValidFlag);
-    void DisplayWarning(QString warningMessage);
-    // Display dependent scattering warning
-    void DisplayScatteringRegimeWarning(double clearanceToWavelength, double sizeParameter,
+    void PrepareScatteringRegimeWarning(double clearanceToWavelength, double sizeParameter,
                                         double volFraction, double criticalWavelength,
                                         QString strRegime);
+    void DisplayWarning(QString warningMessage);
 
 private:
     Calculate *mCalc;
