+++ paper/paper.bib
@@ -96,17 +96,6 @@
 	doi          = {10.1016/j.jqsrt.2009.03.001},
 	issn         = {00224073}
 }
-@article{Ivezic1996,
-	title = {An investigation of dependent/independent scattering regimes using a discrete dipole approximation},
-	journal = {International Journal of Heat and Mass Transfer},
-	author = {Zˇ. Ivezić and M. Pinar Mengüç},
-	volume = {39},
-	number = {4},
-	pages = {811-822},
-	year = {1996},
-	issn = {0017-9310},
-	doi = {10.1016/0017-9310(95)00142-5},
-}
 @article{Jacques2013,
 	title        = {{Optical properties of biological tissues: a review}},
 	author       = {Jacques, Steven L.},
