+++ paper/paper.md
@@ -32,7 +32,7 @@ affiliations:
       Beckman Laser Institute and Medical Clinic, University of California,
       Irvine, California 92697, United States
     index: 2
-date: 2 February 2026
+date: 4 February 2026
 bibliography: paper.bib
 
 ---
@@ -41,7 +41,7 @@ bibliography: paper.bib
 
 # Summary
 
-Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Its predictions are essential across various scientific and engineering disciplines, including biomedical optics, atmospheric optics, particle characterization, nanofluids, computer graphics, and remote sensing. We developed `MieSimulatorGUI` to bridge the gap for researchers who require Mie simulations but lack specialized programming expertise. By integrating a high-performance C/C++ computational engine with the Qt framework, this user-friendly cross-platform tool calculates key optical properties such as scattering coefficients, cross-sections, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. The graphical user interface (GUI) features six interactive panels that allow users to specify input parameters, visualize distributions, and analyze scattering phenomena. Furthermore, it enables the fitting of spectrally varying reduced scattering coefficient parameters, a feature particularly valuable in fields like tissue optics.
+Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Its predictions are essential across various scientific and engineering disciplines, including biomedical optics, atmospheric optics, particle characterization, nanofluids, computer graphics, and remote sensing. We developed `MieSimulatorGUI` to bridge the gap for researchers who require Mie simulations but lack specialized programming expertise. By integrating a high-performance C/C++ computational engine with the Qt framework, this user-friendly cross-platform tool calculates key optical properties such as scattering coefficients, cross-sections, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. The graphical user interface (GUI) features six interactive panels that allow users to specify optical and particle input parameters, visualize particle distributions, and compute scattering metrics. Furthermore, it enables the fitting of the spectral dependence of the reduced scattering coefficient, a feature particularly valuable in fields like tissue optics.
 `MieSimulatorGUI` is open source, hosted on [GitHub](https://github.com/VirtualPhotonics/MieSimulatorGUI), and distributed under [the MIT license](https://opensource.org/license/mit). It is available via the [project's download page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Downloads).
 
 # Statement of need
@@ -50,11 +50,11 @@ Mie theory is derived from Maxwell's equations and provides a comprehensive fram
 
[…]
