+++ paper/paper.md
@@ -41,8 +41,9 @@ bibliography: paper.bib
 
 # Summary
 
-Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Its predictions are essential across various scientific and engineering disciplines, including biomedical optics, atmospheric optics, particle characterization, nanofluids, computer graphics, and remote sensing. We developed `MieSimulatorGUI` to bridge the gap for researchers who require Mie simulations but lack specialized programming expertise. By integrating a high-performance C/C++ computational engine with the Qt framework, this user-friendly cross-platform tool calculates key optical properties such as scattering coefficients, cross-sections, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. The graphical user interface (GUI) features six interactive panels that allow users to specify optical and particle input parameters, visualize particle distributions, and compute scattering metrics. Furthermore, it enables the fitting of the spectral dependence of the reduced scattering coefficient, a feature particularly valuable in fields like tissue optics.
-`MieSimulatorGUI` is open source, hosted on [GitHub](https://github.com/VirtualPhotonics/MieSimulatorGUI), and distributed under [the MIT license](https://opensource.org/license/mit). It is available via the [project's download page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Downloads).
+Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Its predictions are essential across various scientific and engineering disciplines, including biomedical optics, atmospheric optics, particle characterization, nanofluids, computer graphics, and remote sensing. We developed `MieSimulatorGUI` to bridge the gap for researchers who require Mie simulations but lack specialized programming expertise. By integrating a high-performance C/C++ computational engine with the Qt framework, this user-friendly cross-platform tool calculates key optical properties such as scattering coefficients, cross-sections, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. The graphical user interface (GUI) features six interactive panels that allow users to specify optical and particle input parameters, visualize particle distributions, and compute scattering metrics. Furthermore, it enables the fitting of the spectral dependence of the reduced scattering coefficient, a feature particularly valuable in fields like tissue optics. Licensed under MIT, `MieSimulatorGUI` is an open-source project hosted on [GitHub](https://github.com/VirtualPhotonics/MieSimulatorGUI) and accessible via its [download page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Downloads).
+
+
 
 # Statement of need
 
@@ -75,7 +76,7 @@ The tool provides options for either monodisperse (uniform-sized) or polydispers
 
 ## Number Density Panel
 
-This panel graphically presents the number density of spheres $N_s$ [\#/ $\text{mm}^3$] used in the simulation.  The subsequent tab displays the `Size Parameter` defined as 2$\pi Rn_{med} / \lambda_{vacuum}$ [@Bohren1983], where $R$ [ $\mu\text{m}$ ] denotes the particle radius, $n_{med}$  the medium's refractive index, and $\lambda_{vacuum}$ [ $\mu\text{m}$ ] is the wavelength in vacuum. 
[…]
