### 2026-02-05 3ea7c3e
Improved dependent scattering warning message
