+++ src/test/test_calculate.cpp
@@ -482,7 +482,7 @@ void TestCalculate::test_ComputeMuspAtRefWavel_sanityCheck()
 // Test case: High concentration regime (f_v > 0.1)
 void TestCalculate::test_CheckIndependentScattering_HighConcentration()
 {
-    double clearanceToWavelength, sizeParameter, volFraction, criticalWavelength;
+    double clearanceToWavelength, sizeParameter, volFraction, wavelength, clearance;
     QString strRegime;
     bool flagVolFrac = false;
 
@@ -493,15 +493,15 @@ void TestCalculate::test_CheckIndependentScattering_HighConcentration()
     mPara->medRef = 1.0;
 
     bool isDependent = mCalc->CheckIndependentScattering(mPara, clearanceToWavelength, sizeParameter,
-                                                         volFraction, criticalWavelength, strRegime,
-                                                         flagVolFrac);
+                                                         volFraction, wavelength, clearance,
+                                                         strRegime, flagVolFrac);
     QVERIFY(isDependent == true);
 }
 
 // Test case: Low concentration, size parameter below threshold (f_v <= 0.006, sizePara <= 0.388)
 void TestCalculate::test_CheckIndependentScattering_LowConcSmallSize()
 {
-    double clearanceToWavelength, sizeParameter, volFraction, criticalWavelength;
+    double clearanceToWavelength, sizeParameter, volFraction, wavelength, clearance;
     QString strRegime;
     bool flagVolFrac = false;
 
@@ -512,15 +512,15 @@ void TestCalculate::test_CheckIndependentScattering_LowConcSmallSize()
     mPara->medRef = 1.0;
 
     bool isDependent = mCalc->CheckIndependentScattering(mPara, clearanceToWavelength, sizeParameter,
-                                                         volFraction, criticalWavelength, strRegime,
-                                                         flagVolFrac);
+                                                         volFraction, wavelength, clearance,
+                                                         strRegime, flagVolFrac);
     QVERIFY(isDependent == false);
 }
 
 // Test case: Transitional regime, Dependent (0.006 < f_v <= 0.1 and low clearance)
 void TestCalculate::test_CheckIndependentScattering_TransitionalDependent()
 {
-    double clearanceToWavelength, sizeParameter, volFraction, criticalWavelength;
+    double clearanceToWavelength, sizeParameter, volFraction, wavelength, clearance;
     QString strRegime;
     bool flagVolFrac = false;
 
@@ -530,19 +530,19 @@ void TestCalculate::test_CheckIndependentScattering_TransitionalDependent()
     mPara->numDensityArray[0] = 5e6;
     mPara->medRef = 1.0;
 
-    // interParticleDistance = 1000 / (5e6)^(1/3) = 5.84
+    // clearance = 1000 / (5e6)^(1/3) = 5.84
     // clearanceToWavelength = (5.84 - 2*1.0) / 1.0 = 3.84
     // 3.84 <= 5.0 (required clearance)
     bool isDependent = mCalc->CheckIndependentScattering(mPara, clearanceToWavelength, sizeParameter,
-                                                         volFraction, criticalWavelength, strRegime,
-                                                         flagVolFrac);
+                                                         volFraction, wavelength, clearance,
+                                                         strRegime, flagVolFrac);
     QVERIFY(isDependent == true);
 }
 
 // Test case: Low concentration, size parameter above threshold, Independent (High clearance)
 void TestCalculate::test_CheckIndependentScattering_LowConcLargeSizeIndependent()
 {
-    double clearanceToWavelength, sizeParameter, volFraction, criticalWavelength;
[…]
