### 2026-02-05 2568627
Fixed font color of warning message
