+++ paper/paper.md
@@ -25,12 +25,10 @@ authors:
     affiliation: '1, 2'
 affiliations:
   - name: >-
-      Department of Chemical and Biomolecular Engineering, University of
-      California, Irvine, California 92697, United States
+      Department of Chemical and Biomolecular Engineering, University of California, Irvine, California 92697, United States
     index: 1
   - name: >-
-      Beckman Laser Institute and Medical Clinic, University of California,
-      Irvine, California 92697, United States
+      Beckman Laser Institute and Medical Clinic, University of California, Irvine, California 92697, United States
     index: 2
 date: 4 February 2026
 bibliography: paper.bib
@@ -43,11 +41,9 @@ bibliography: paper.bib
 
 Mie theory is a mathematical framework derived from Maxwell's equations that models electromagnetic scattering by spherical dielectric particles.  Its predictions are essential across various scientific and engineering disciplines, including biomedical optics, atmospheric optics, particle characterization, nanofluids, computer graphics, and remote sensing. We developed `MieSimulatorGUI` to bridge the gap for researchers who require Mie simulations but lack specialized programming expertise. By integrating a high-performance C/C++ computational engine with the Qt framework, this user-friendly cross-platform tool calculates key optical properties such as scattering coefficients, cross-sections, angular scattering distributions, and scattering asymmetry for both monodisperse and polydisperse particle distributions. The graphical user interface (GUI) features six interactive panels that allow users to specify optical and particle input parameters, visualize particle distributions, and compute scattering metrics. Furthermore, it enables the fitting of the spectral dependence of the reduced scattering coefficient, a feature particularly valuable in fields like tissue optics. Licensed under MIT, `MieSimulatorGUI` is an open-source project hosted on [GitHub](https://github.com/VirtualPhotonics/MieSimulatorGUI) and accessible via its [download page](https://github.com/VirtualPhotonics/MieSimulatorGUI/wiki/Downloads).
 
-
-
 # Statement of need
 
-Mie theory is derived from Maxwell's equations and provides a comprehensive framework for modeling electromagnetic scattering by spherical particles [@Mie1908]. Mie theory is utilized across diverse fields, ranging from nanomaterials and biomedical optics to atmospheric science and astronomy [@Goody1989; @Saidi1995; @Wang2005; @Chalut2008; @Horvath2009; @Bhandari11]. Despite its broad applicability, the theory's reliance on complex mathematical constructs, such as infinite series and special functions [@VandeHulst1957; @Bohren1983; @Wiscombe1980; @Majic2020], demands advanced computational implementation. 
+Mie theory is derived from Maxwell's equations and provides a comprehensive framework for modeling electromagnetic scattering by spherical particles [@Horvath2009; @Mie1908]. Mie theory is utilized across diverse fields, ranging from nanomaterials and biomedical optics to atmospheric science and astronomy [@Goody1989; @Saidi1995; @Wang2005; @Chalut2008; @Bhandari11]. Despite its broad applicability, the theory's reliance on complex mathematical constructs, such as infinite series and special functions [@VandeHulst1957; @Bohren1983; @Wiscombe1980; @Majic2020], demands advanced computational implementation. 
 
[…]
