### 2026-02-05 #9551 @janakarana
Dear @finsberg

We have addressed the [issue](https://github.com/VirtualPhotonics/MieSimulatorGUI/issues/12) you raised regarding the documentation. We also created a new release, v2.0, and archived it on Zenodo. The DOI for this release is [10.5281/zenodo.18498015](https://doi.org/10.5281/zenodo.18462320). As requested, the archived version maintains the same title, authors, and license as the repository. The latest version of the paper has been updated and is ready for your review.
