### 2026-02-06 13a9473
Remove invalid isbn from bib file
