+++ paper/paper.bib
@@ -219,7 +219,6 @@
 	number       = 13,
 	pages        = {2788--97},
 	doi          = {10.1364/AO.37.002788},
-	isbn         = {0003-6935},
 	issn         = {0003-6935},
 	url          = {http://www.ncbi.nlm.nih.gov/pubmed/18273225},
 	pmid         = 18273225
