### 2026-02-06 8b5e4d7
Removed  erronous isbn
