+++ paper/paper.bib
@@ -217,10 +217,8 @@
 	journal      = {Applied optics},
 	volume       = 37,
 	number       = 13,
-	pages        = {2788--97},
+	pages        = {2788--2797},
 	doi          = {10.1364/AO.37.002788},
-	isbn         = {0003-6935},
-	issn         = {0003-6935},
 	url          = {http://www.ncbi.nlm.nih.gov/pubmed/18273225},
 	pmid         = 18273225
 }
