### 2026-02-06 #9551 @janakarana
Dear @finsberg

Removed isbn.
