+++ README.MD
@@ -33,7 +33,7 @@ Use the following citation or acknowledgement in publications or applications th
 
 In addition, for any work that utilizes the Mie Simulator GUI software, cite the following publication"
 
-__Janaka C. Ranasinghesagara, Carole K. Hayakawa, Lisa Malenfant, Vasan Venugopalan, "MieSimulatorGUI: A user-friendly tool to compute and visualize light scattering by spherical dielectric particles," Journal of Open Source Software 11(118), 9551(1-6) (2026) https://doi.org/10.21105/joss.09551__
+__Janaka C. Ranasinghesagara, Carole K. Hayakawa, Lisa M. Glover, Vasan Venugopalan, "MieSimulatorGUI: A user-friendly tool to compute and visualize light scattering by spherical dielectric particles," Journal of Open Source Software 11(118), 9551 (2026) https://doi.org/10.21105/joss.09551__
 
 ## Software
 _Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:_
@@ -46,6 +46,7 @@ _THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPL
 Virtual Tissue Simulator (VTS) Project.
 
 Copyright ©2026 [Virtual Photonics Technology Initiative](https://virtualphotonics.org/).
+
 
 
 
