+++ paper/paper.bib
@@ -8,19 +8,20 @@
 	publisher = {Optica Publishing Group},
 	title = {Supercontinuum laser based optical characterization of Intralipid{\textregistered} phantoms in the 500-2250 nm range},
 	volume = {21},
+	number = {26},
 	month = {Dec},
 	year = {2013},
 	url = {https://opg.optica.org/oe/abstract.cfm?URI=oe-21-26-32450},
 	doi = {10.1364/OE.21.032450},
 }
-@article{Bhandari11,
+@article{Bhandari2011,
 	title      = {Modeling optical properties of human skin using Mie theory for particles with different size distributions and refractive indices},
 	author     = {A. Bhandari and B. Hamre and {\O}. Frette and K. Stamnes and J. J. Stamnes},
 	journal    = {Opt. Express},	
+	volume     = {19},
 	number     = {15},
 	pages      = {14549--14567},
-	publisher  = {Optica Publishing Group},	
-	volume     = {19},
+	publisher  = {Optica Publishing Group},		
 	month = {Jul},
 	year = {2011},
 	doi = {10.1364/OE.19.014549}
@@ -35,18 +36,19 @@
 @article{Chalut2008,
 	title={Application of Mie theory to assess structure of spheroidal scattering in backscattering geometries},
 	author={Chalut, Kevin J and Giacomelli, Michael G and Wax, Adam},
-	journal={J Opt Soc Am A Opt Image Sci Vis},
+	journal={J Opt Soc Am A},
 	volume={25},
 	number={8},
 	pages={1866--1874},
 	year={2008},
+	month = {Aug},
 	publisher={Optical Society of America},
 	doi={10.1364/josaa.25.001866}
 }
 @article{DiNinni2011,
 	author = {Di Ninni, P and Martelli, F and Zaccanti, G},
 	title = {{Intralipid: towards a diffusive reference standard for optical tissue phantoms}},
-	journal = {Physics in Medicine \& Biology},
+	journal = {Phys. Med. Biol.},
 	volume = {56},
 	number = {2},
 	pages = {N21--N28},
@@ -57,7 +59,7 @@
 @article{Galy2020,
 	title = {Revisiting independent versus dependent scattering regimes in suspensions or aggregates of spherical particles},
 	author = {Galy, Tiphaine and Huang, Daniel and Pilon, Laurent},
-	journal = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+	journal = {J. Quant. Spectrosc. Radiat. Transfer},
 	volume = {246},
 	pages = {106924},
 	year = {2020},
@@ -69,7 +71,7 @@
 	title        = {{Phase function simulation in tissue phantoms: a fractal approach}},
 	author       = {G{\'{e}}l{\'{e}}bart, Bernard and Tinet, Eric and Tualle, Jean Michel and Avrillier, Sigrid},
 	year         = 1996,
-	journal      = {Pure and Applied Optics},
+	journal      = {Pure and Appl. Opt.},
 	volume       = 5,
 	number       = 4,
 	pages        = {377--388},
@@ -86,37 +88,35 @@
 	edition      = {2nd Edition}
 }
 @article{Horvath2009,
-	title        = {{Light scattering: Mie and More commemorating 100 years of Mie's 1908 publication}},
+	title        = {Light scattering: Mie and more – Commemorating 100 years of Mie's 1908 publication},
 	author       = {Horvath, Helmuth},
 	year         = 2009,
-	journal      = {Journal of Quantitative Spectroscopy and Radiative Transfer},
+	journal      = {J. Quant. Spectrosc. Radiat. Transfer},
 	volume       = 110,
 	number       = 11,
 	pages        = {783--786},
[…]
