### 2026-02-08 3aef289
Fixed bibtex
