+++ paper/paper.bib
@@ -32,6 +32,7 @@
 	year         = 1983,
 	publisher    = {John Wiley and Sons, Inc.},
 	address      = {New York}
+	isbn         = {047105772X}
 }
 @article{Chalut2008,
 	title={Application of Mie theory to assess structure of spheroidal scattering in backscattering geometries},
@@ -81,11 +82,12 @@
 @book{Goody1989,
 	title        = {{Atmospheric Radiation: Theoretical Basis}},
 	author       = {Goody, R. M. and Yung, Y. L.},
-	year         = 1989,
+	year         = 1995,
 	publisher    = {Oxford University Press},
 	address      = {Oxford},
-	pages        = 544,
+	pages        = 540,
 	edition      = {2nd Edition}
+	issn         = {978-0195356106}
 }
 @article{Horvath2009,
 	title        = {Light scattering: Mie and more – Commemorating 100 years of Mie's 1908 publication},
@@ -152,10 +154,9 @@
 @article{Mourant1997,
 	author = {Judith R. Mourant and Tamika Fuselier and James Boyer and Tamara M. Johnson and Irving J. Bigio},
 	journal = {Appl. Opt.},
-	keywords = {Absorption coefficient; Elastic scattering; Mie scattering; Optical properties; Refractive index; Turbid media},
 	pages = {949--957},
 	publisher = {Optica Publishing Group},
-	title = {Predictions and measurements of scattering and absorption over broadwavelength ranges in tissue phantoms},
+	title = {Predictions and measurements of scattering and absorption over broad wavelength ranges in tissue phantoms},
 	volume = {36},
 	number = {4},
 	month = {Feb},
@@ -217,13 +218,13 @@
 	doi          = {10.1364/AO.37.002788},
 	issn         = {0003-6935}
 }
-@book{Seinfeld1998,
+@book{Seinfeld1997,
 	author = {Seinfeld, John H. and Pandis, Spyros N.},
-	title = {Atmospheric Chemistry and Physics: From Air Pollution to Climate Change},
-	publisher = {Wiley},
+	title = {Atmospheric Chemistry and Physics: Air Pollution to Climate Change},
+	publisher = {Wiley-Interscience},
 	address = {New York},
-	year = {1998},
-	isbn = {0-471-17815-2}
+	year = {1997},
+	isbn = {0471178160}
 }
 @article{Sumlin2018,
 	title        = {Retrieving the aerosol complex refractive index using PyMieScatt: A Mie computational package with visualization capabilities},
@@ -257,7 +258,7 @@
 }
 @book{VandeHulst1957,
 	title        = {{Light scattering by small particles}},
-	author       = {van de Hulst, H. C.},
+	author       = {H. C. van de Hulst},
 	year         = 1957,
 	publisher    = {John Wiley and Sons Inc.},
 	address      = {New York},
