+++ paper/paper.md
@@ -47,7 +47,7 @@ Mie theory is derived from Maxwell's equations and provides a comprehensive fram
 
 While numerous Mie simulation packages are available (many of which are listed on [SCATTPORT.org](https://scattport.org/index.php/light-scattering-software) and [Wikipedia](https://en.wikipedia.org/wiki/Codes_for_electromagnetic_scattering_by_spheres)), they generally fall into two categories: older, established codes focusing on computational efficiency [@Wiscombe1980; @Bohren1983], and newer, object-oriented libraries typically hosted on version-control platforms [@Sumlin2018; @PoinsinetdeSivry-Houle2023; @Prahl_mie; @MieScattering]. Although both categories provide robust computational engines, they usually demand significant programming proficiency. This requirement creates a barrier for experimentalists, clinical scientists, and educators who need these analytical capabilities but may lack the specialized coding expertise to integrate such libraries into their workflows.
 
-`MieSimulatorGUI` bridges this gap by providing an intuitive, cross-platform desktop application that computes and fits scattering parameters for monodisperse and polydisperse distributions without any coding. Unlike standard implementations, it supports heterogeneous polydispersity, allowing users to assign bin-specific complex refractive indices via custom data inputs, a feature often absent in simplified GUI tools. The tool facilitates high-impact use cases such as biomedical optics [@Mourant1997; @Wang2005; @Jacques2013] and atmospheric research [@Seinfeld1998; @Teri2022], where users can define complex particle configurations and directly fit spectrally-varying reduced scattering coefficients. By integrating a powerful C/C++ computational engine with intuitive [Qt](https://www.qt.io/) interface, `MieSimulatorGUI` offers accessible, yet powerful Mie theory computations, facilitating both streamlined research analysis and interactive pedagogical demonstrations. 
+`MieSimulatorGUI` bridges this gap by providing an intuitive, cross-platform desktop application that computes and fits scattering parameters for monodisperse and polydisperse distributions without any coding. Unlike standard implementations, it supports heterogeneous polydispersity, allowing users to assign bin-specific complex refractive indices via custom data inputs, a feature often absent in simplified GUI tools. The tool facilitates high-impact use cases such as biomedical optics [@Mourant1997; @Wang2005; @Jacques2013] and atmospheric research [@Seinfeld1997; @Teri2022], where users can define complex particle configurations and directly fit spectrally-varying reduced scattering coefficients. By integrating a powerful C/C++ computational engine with intuitive [Qt](https://www.qt.io/) interface, `MieSimulatorGUI` offers accessible, yet powerful Mie theory computations, facilitating both streamlined research analysis and interactive pedagogical demonstrations. 
 
 # Main Features
 
