+++ paper/paper.bib
@@ -10,9 +10,7 @@
 	volume = {21},
 	number = {26},
 	month = {Dec},
-	year = {2013},
-	url = {https://opg.optica.org/oe/abstract.cfm?URI=oe-21-26-32450},
-	doi = {10.1364/OE.21.032450},
+	year = {2013}
 }
 @article{Bhandari2011,
 	title      = {Modeling optical properties of human skin using Mie theory for particles with different size distributions and refractive indices},
@@ -31,7 +29,7 @@
 	author       = {Bohren, C F and Huffman, D R},
 	year         = 1983,
 	publisher    = {John Wiley and Sons, Inc.},
-	address      = {New York}
+	address      = {New York},
 	isbn         = {047105772X}
 }
 @article{Chalut2008,
@@ -55,7 +53,7 @@
 	pages = {N21--N28},
 	year = {2011},
 	publisher = {IOP Publishing},
-	doi = {10.1088/0031-9155/56/2/N01},  
+	doi = {10.1088/0031-9155/56/2/N01}  
 }
 @article{Galy2020,
 	title = {Revisiting independent versus dependent scattering regimes in suspensions or aggregates of spherical particles},
