+++ paper/paper.bib
@@ -84,8 +84,8 @@
 	publisher    = {Oxford University Press},
 	address      = {Oxford},
 	pages        = 540,
-	edition      = {2nd Edition}
-	issn         = {978-0195356106}
+	edition      = {2nd Edition},
+	isbn         = {978-0195356106}
 }
 @article{Horvath2009,
 	title        = {Light scattering: Mie and more – Commemorating 100 years of Mie's 1908 publication},
@@ -235,7 +235,7 @@
 	doi          = {10.1016/j.jqsrt.2017.10.012},
 	publisher    = {Elsevier}
 }
-@Article{Teri2022,
+@article{Teri2022,
 	author = {Teri, M. and M\"uller, T. and Gasteiger, J. and Valentini, S. and Horvath, H. and Vecchi, R. and Bauer, P. and Walser, A. and Weinzierl, B.},
 	title = {Impact of particle size, refractive index, and shape on the determination of the particle scattering coefficient – an optical closure study evaluating different nephelometer angular truncation and illumination corrections},
 	journal = {Atmosph. Meas. Tech.},
