### 2026-02-09 #9551 @janakarana
Dear @finsberg

Thank you for your inquiry. During the proofreading, we identified several clerical errors in the reference list, including incorrect DOIs, volume numbers, isbn and journal naming formats. We updated these to ensure bibliographic accuracy and metadata integrity.

We would like to clarify that no changes were made to the main text of the manuscript. The multiple PDF generations were due to minor BibTeX synchronization issues encountered while fixing these citations. Our GitHub repository reflects these bibliographic fixes. We apologize for not explicitly mentioning these bibliographic corrections immediately upon submission.
