### 2026-02-10 b95b42b
Add Open Journals PDF generator
