### 2026-02-10 1896cfc
Fixed bibTex for van de Hulst
