+++ paper/paper.bib
@@ -256,7 +256,7 @@
 }
 @book{VandeHulst1957,
 	title        = {{Light scattering by small particles}},
-	author       = {H. C. van de Hulst},
+	author       = {H. C. van~de~Hulst},
 	year         = 1957,
 	publisher    = {John Wiley and Sons Inc.},
 	address      = {New York},
