+++ paper/paper.bib
@@ -162,7 +162,7 @@
 	doi = {10.1364/AO.36.000949}
 }
 @article{PoinsinetdeSivry-Houle2023,
-    author = {Martin Poinsinet de Sivry-Houle and Nicolas Godbout and Caroline Boudoux},
+    author = {Martin {Poinsinet~de~Sivry-Houle} and Nicolas Godbout and Caroline Boudoux},
     journal = {Opt. Continuum},
     title = {PyMieSim: an open-source library for fast and flexible far-field Mie scattering simulations},
     volume = {2},
@@ -256,14 +256,14 @@
 }
 @book{VandeHulst1957,
 	title        = {{Light scattering by small particles}},
-	author       = {H. C. van~de~Hulst},
+	author       = {H. C. {van~de~Hulst}},
 	year         = 1957,
 	publisher    = {John Wiley and Sons Inc.},
 	address      = {New York},
 	pages        = 470
 }
 @article{vanStaveren1991,
-	author  = {Hugo J. van Staveren and Christian J. M. Moes and Jan van Marie and Scott A. Prahl and Martin J. C. van Gemert},
+	author  = {Hugo J. {van~Staveren} and Christian J. M. Moes and Jan van Marie and Scott A. Prahl and Martin J. C. van Gemert},
 	title   = {Light scattering in lntralipid-10\% in the wavelength range of 400–1100 nm},
 	journal = {Appl. Opt.},
 	year    = {1991},
@@ -285,7 +285,7 @@
 }
 @techreport{Wiscombe1979,
 	title        = {{Mie Scattering Calculations : Advances in Technique and Fast, Vector-Speed Computer Codes}},
-	author       = {Wiscombe, Warren J},
+	author       = {W. J. Wiscombe},
 	year={1979},
 	institution={National Center for Atmospheric Research (NCAR)},
 	address={Boulder, Colorado},
