### 2026-02-10 ee0c414
fix bibTex for vanStaveren et al.
