### 2026-02-10 7f39a65
fix bibTex for vanStaveran et al.
