+++ paper/paper.bib
@@ -263,7 +263,7 @@
 	pages        = 470
 }
 @article{vanStaveren1991,
-	author  = {van Staveren, H. J. and C. J. M. Moes and J. van Marle and S. A. Prahl and M. J. C. van Gemert},
+	author  = {{van Staveren}, H. J. and C. J. M. Moes and J. van Marle and S. A. Prahl and M. J. C. van Gemert},
 	title   = {Light scattering in lntralipid-10\% in the wavelength range of 400–1100 nm},
 	journal = {Appl. Opt.},
 	year    = {1991},
