### 2026-02-10 cb1a81c
Fix bibliography
