### 2026-02-10 a18c39a
fix bibliography
