+++ paper/paper.bib
@@ -162,7 +162,7 @@
 	doi = {10.1364/AO.36.000949}
 }
 @article{PoinsinetdeSivry-Houle2023,
-    author = {Martin Poinsinet {de~Sivry-Houle} and Nicolas Godbout and Caroline Boudoux},
+    author = {{de~Sivry-Houle}, Martin Poinsinet  and Godbout, Nicolas  and Boudoux, Caroline},
     journal = {Opt. Continuum},
     title = {PyMieSim: an open-source library for fast and flexible far-field Mie scattering simulations},
     volume = {2},
