### 2026-02-10 #9551 @janakarana
Dear @finsberg

While I previously believed all issues were resolved, one of our member identified a formatting error regarding multi-word surnames in the bibliography.

We realized that surnames consisting of two or three words were not being handled correctly by the JOSS bibliography renderer. Since JOSS requires specific formatting for these cases, we have updated the .bib file to ensure they are displayed accurately. For example, "van Staveren, H. J.".in bib file was previously rendered as "Staveren, H. J. van". we have implemented stricter formatting using LaTeX braces and ~, such as {van ~Staveren}, H. J" to ensure proper rendering.

The above PDF is our final version and is ready for formal acceptance. Thank you for your patience.
